refactor: move Go proto generated code into backend module
Consolidate proto/gen/go (separate Go module) into backend/gen/proto
to simplify the workspace and eliminate the external module dependency.
Update all import paths, buf.gen.yaml output config, and Makefile targets.
