+++ backend/cmd/taskguild-agent/main.go
@@ -15,8 +15,8 @@ import (
 	"time"
 
 	"connectrpc.com/connect"
-	v1 "github.com/kazz187/taskguild/proto/gen/go/taskguild/v1"
-	"github.com/kazz187/taskguild/proto/gen/go/taskguild/v1/taskguildv1connect"
+	v1 "github.com/kazz187/taskguild/backend/gen/proto/taskguild/v1"
+	"github.com/kazz187/taskguild/backend/gen/proto/taskguild/v1/taskguildv1connect"
 	"github.com/oklog/ulid/v2"
 )
