+++ backend/cmd/taskguild-agent/sync.go
@@ -9,8 +9,8 @@ import (
 	"strings"
 
 	"connectrpc.com/connect"
-	v1 "github.com/kazz187/taskguild/proto/gen/go/taskguild/v1"
-	"github.com/kazz187/taskguild/proto/gen/go/taskguild/v1/taskguildv1connect"
+	v1 "github.com/kazz187/taskguild/backend/gen/proto/taskguild/v1"
+	"github.com/kazz187/taskguild/backend/gen/proto/taskguild/v1/taskguildv1connect"
 )
 
 // syncAgents calls the SyncAgents RPC and writes .claude/agents/*.md files locally.
