+++ backend/cmd/taskguild-server/main.go
@@ -33,7 +33,7 @@ import (
 	"github.com/kazz187/taskguild/backend/pkg/clog"
 	"github.com/kazz187/taskguild/backend/pkg/storage"
 
-	taskguildv1 "github.com/kazz187/taskguild/proto/gen/go/taskguild/v1"
+	taskguildv1 "github.com/kazz187/taskguild/backend/gen/proto/taskguild/v1"
 
 	server "github.com/kazz187/taskguild/backend/internal"
 )
