+++ backend/go.mod
@@ -13,7 +13,6 @@ require (
 	github.com/fatih/color v1.15.0
 	github.com/go-chi/chi/v5 v5.2.2
 	github.com/kazz187/claude-agent-sdk-go v0.0.10
-	github.com/kazz187/taskguild/proto/gen/go v0.0.7
 	github.com/kelseyhightower/envconfig v1.4.0
 	github.com/oklog/ulid/v2 v2.1.1
 	github.com/rs/cors v1.11.1
