+++ backend/internal/event/server.go
@@ -6,8 +6,8 @@ import (
 	"connectrpc.com/connect"
 
 	"github.com/kazz187/taskguild/backend/internal/eventbus"
-	taskguildv1 "github.com/kazz187/taskguild/proto/gen/go/taskguild/v1"
-	"github.com/kazz187/taskguild/proto/gen/go/taskguild/v1/taskguildv1connect"
+	taskguildv1 "github.com/kazz187/taskguild/backend/gen/proto/taskguild/v1"
+	"github.com/kazz187/taskguild/backend/gen/proto/taskguild/v1/taskguildv1connect"
 )
 
 var _ taskguildv1connect.EventServiceHandler = (*Server)(nil)
