+++ backend/internal/orchestrator/orchestrator.go
@@ -10,7 +10,7 @@ import (
 	"github.com/kazz187/taskguild/backend/internal/project"
 	"github.com/kazz187/taskguild/backend/internal/task"
 	"github.com/kazz187/taskguild/backend/internal/workflow"
-	taskguildv1 "github.com/kazz187/taskguild/proto/gen/go/taskguild/v1"
+	taskguildv1 "github.com/kazz187/taskguild/backend/gen/proto/taskguild/v1"
 )
 
 type Orchestrator struct {
