+++ go.work
@@ -1,6 +1,3 @@
 go 1.26.0
 
-use (
-	./backend
-	./proto/gen/go
-)
+use ./backend
