+++ proto/Makefile
@@ -2,13 +2,13 @@
 
 generate:
 	buf generate
-	cd gen/go && go mod tidy
+	cd ../backend && go mod tidy
 
 lint:
 	buf lint
 
 tidy:
-	cd gen/go && go mod tidy
+	cd ../backend && go mod tidy
 
 clean:
-	rm -rf gen/go/taskguild gen/go/buf gen/ts
+	rm -rf ../backend/gen/proto/taskguild ../backend/gen/proto/buf gen/ts
