--- proto/gen/go/go.mod
@@ -1,8 +1,0 @@
-module github.com/kazz187/taskguild/proto/gen/go
-
-go 1.26.0
-
-require (
-	connectrpc.com/connect v1.18.1
-	google.golang.org/protobuf v1.36.7
-)
