+++ proto/gen/ts/taskguild/v1/agent_manager_pb.ts
@@ -18,7 +18,7 @@ import type { Message } from "@bufbuild/protobuf";
  * Describes the file taskguild/v1/agent_manager.proto.
  */
 export const file_taskguild_v1_agent_manager: GenFile = /*@__PURE__*/
-  fileDesc("CiB0YXNrZ3VpbGQvdjEvYWdlbnRfbWFuYWdlci5wcm90bxIMdGFza2d1aWxkLnYxImwKHEFnZW50TWFuYWdlclN1YnNjcmliZVJlcXVlc3QSGAoQYWdlbnRfbWFuYWdlcl9pZBgBIAEoCRIUCgxwcm9qZWN0X25hbWUYAiABKAkSHAoUbWF4X2NvbmN1cnJlbnRfdGFza3MYAyABKAUiyQIKDEFnZW50Q29tbWFuZBI8Cg50YXNrX2F2YWlsYWJsZRgBIAEoCzIiLnRhc2tndWlsZC52MS5UYXNrQXZhaWxhYmxlQ29tbWFuZEgAEjYKC2Fzc2lnbl90YXNrGAIgASgLMh8udGFza2d1aWxkLnYxLkFzc2lnblRhc2tDb21tYW5kSAASNgoLY2FuY2VsX3Rhc2sYAyABKAsyHy50YXNrZ3VpbGQudjEuQ2FuY2VsVGFza0NvbW1hbmRIABJIChRpbnRlcmFjdGlvbl9yZXNwb25zZRgEIAEoCzIoLnRhc2tndWlsZC52MS5JbnRlcmFjdGlvblJlc3BvbnNlQ29tbWFuZEgAEjYKC3N5bmNfYWdlbnRzGAUgASgLMh8udGFza2d1aWxkLnYxLlN5bmNBZ2VudHNDb21tYW5kSABCCQoHY29tbWFuZCLEAQoUVGFza0F2YWlsYWJsZUNvbW1hbmQSDwoHdGFza19pZBgBIAEoCRINCgV0aXRsZRgCIAEoCRIXCg9hZ2VudF9jb25maWdfaWQYAyABKAkSQgoIbWV0YWRhdGEYBCADKAsyMC50YXNrZ3VpbGQudjEuVGFza0F2YWlsYWJsZUNvbW1hbmQuTWV0YWRhdGFFbnRyeRovCg1NZXRhZGF0YUVudHJ5EgsKA2tleRgBIAEoCRINCgV2YWx1ZRgCIAEoCToCOAEi3gEKEUFzc2lnblRhc2tDb21tYW5kEg8KB3Rhc2tfaWQYASABKAkSFwoPYWdlbnRfY29uZmlnX2lkGAIgASgJEhQKDGluc3RydWN0aW9ucxgDIAEoCRIXCg93b3JrdHJlZV9icmFuY2gYBCABKAkSPwoIbWV0YWRhdGEYBSADKAsyLS50YXNrZ3VpbGQudjEuQXNzaWduVGFza0NvbW1hbmQuTWV0YWRhdGFFbnRyeRovCg1NZXRhZGF0YUVudHJ5EgsKA2tleRgBIAEoCRINCgV2YWx1ZRgCIAEoCToCOAEiNAoRQ2FuY2VsVGFza0NvbW1hbmQSDwoHdGFza19pZBgBIAEoCRIOCgZyZWFzb24YAiABKAkiRgoaSW50ZXJhY3Rpb25SZXNwb25zZUNvbW1hbmQSFgoOaW50ZXJhY3Rpb25faWQYASABKAkSEAoIcmVzcG9uc2UYAiABKAkiEwoRU3luY0FnZW50c0NvbW1hbmQiPQoQQ2xhaW1UYXNrUmVxdWVzdBIPCgd0YXNrX2lkGAEgASgJEhgKEGFnZW50X21hbmFnZXJfaWQYAiABKAkixQEKEUNsYWltVGFza1Jlc3BvbnNlEg8KB3N1Y2Nlc3MYASABKAgSFwoPYWdlbnRfY29uZmlnX2lkGAIgASgJEhQKDGluc3RydWN0aW9ucxgDIAEoCRI/CghtZXRhZGF0YRgEIAMoCzItLnRhc2tndWlsZC52MS5DbGFpbVRhc2tSZXNwb25zZS5NZXRhZGF0YUVudHJ5Gi8KDU1ldGFkYXRhRW50cnkSCwoDa2V5GAEgASgJEg0KBXZhbHVlGAIgASgJOgI4ASJgChdSZXBvcnRUYXNrUmVzdWx0UmVxdWVzdBIPCgd0YXNrX2lkGAEgASgJEg8KB3N1bW1hcnkYAyABKAkSFQoNZXJyb3JfbWVzc2FnZRgEIAEoCUoECAIQA1IGc3RhdHVzIhoKGFJlcG9ydFRhc2tSZXN1bHRSZXNwb25zZSKBAQoYUmVwb3J0QWdlbnRTdGF0dXNSZXF1ZXN0EhgKEGFnZW50X21hbmFnZXJfaWQYASABKAkSDwoHdGFza19pZBgCIAEoCRIpCgZzdGF0dXMYAyABKA4yGS50YXNrZ3VpbGQudjEuQWdlbnRTdGF0dXMSDwoHbWVzc2FnZRgEIAEoCSIbChlSZXBvcnRBZ2VudFN0YXR1c1Jlc3BvbnNlInEKEEhlYXJ0YmVhdFJlcXVlc3QSGAoQYWdlbnRfbWFuYWdlcl9pZBgBIAEoCRIUCgxhY3RpdmVfdGFza3MYAiABKAUSLQoJdGltZXN0YW1wGAMgASgLMhouZ29vZ2xlLnByb3RvYnVmLlRpbWVzdGFtcCITChFIZWFydGJlYXRSZXNwb25zZSLAAQoYQ3JlYXRlSW50ZXJhY3Rpb25SZXF1ZXN0Eg8KB3Rhc2tfaWQYASABKAkSEAoIYWdlbnRfaWQYAiABKAkSKwoEdHlwZRgDIAEoDjIdLnRhc2tndWlsZC52MS5JbnRlcmFjdGlvblR5cGUSDQoFdGl0bGUYBCABKAkSEwoLZGVzY3JpcHRpb24YBSABKAkSMAoHb3B0aW9ucxgGIAMoCzIfLnRhc2tndWlsZC52MS5JbnRlcmFjdGlvbk9wdGlvbiJLChlDcmVhdGVJbnRlcmFjdGlvblJlc3BvbnNlEi4KC2ludGVyYWN0aW9uGAEgASgLMhkudGFza2d1aWxkLnYxLkludGVyYWN0aW9uIjcKHUdldEludGVyYWN0aW9uUmVzcG9uc2VSZXF1ZXN0EhYKDmludGVyYWN0aW9uX2lkGAEgASgJIlAKHkdldEludGVyYWN0aW9uUmVzcG9uc2VSZXNwb25zZRIuCgtpbnRlcmFjdGlvbhgBIAEoCzIZLnRhc2tndWlsZC52MS5JbnRlcmFjdGlvbiIpChFTeW5jQWdlbnRzUmVxdWVzdBIUCgxwcm9qZWN0X25hbWUYASABKAkiQwoSU3luY0FnZW50c1Jlc3BvbnNlEi0KBmFnZW50cxgBIAMoCzIdLnRhc2tndWlsZC52MS5BZ2VudERlZmluaXRpb24iiQIKFFJlcG9ydFRhc2tMb2dSZXF1ZXN0Eg8KB3Rhc2tfaWQYASABKAkSKQoFbGV2ZWwYAiABKA4yGi50YXNrZ3VpbGQudjEuVGFza0xvZ0xldmVsEi8KCGNhdGVnb3J5GAMgASgOMh0udGFza2d1aWxkLnYxLlRhc2tMb2dDYXRlZ29yeRIPCgdtZXNzYWdlGAQgASgJEkIKCG1ldGFkYXRhGAUgAygLMjAudGFza2d1aWxkLnYxLlJlcG9ydFRhc2tMb2dSZXF1ZXN0Lk1ldGFkYXRhRW50cnkaLwoNTWV0YWRhdGFFbnRyeRILCgNrZXkYASABKAkSDQoFdmFsdWUYAiABKAk6AjgBIhcKFVJlcG9ydFRhc2tMb2dSZXNwb25zZSqOAQoLQWdlbnRTdGF0dXMSHAoYQUdFTlRfU1RBVFVTX1VOU1BFQ0lGSUVEEAASFQoRQUdFTlRfU1RBVFVTX0lETEUQARIYChRBR0VOVF9TVEFUVVNfUlVOTklORxACEhgKFEFHRU5UX1NUQVRVU19XQUlUSU5HEAMSFgoSQUdFTlRfU1RBVFVTX0VSUk9SEAQy1wYKE0FnZW50TWFuYWdlclNlcnZpY2USVQoJU3Vic2NyaWJlEioudGFza2d1aWxkLnYxLkFnZW50TWFuYWdlclN1YnNjcmliZVJlcXVlc3QaGi50YXNrZ3VpbGQudjEuQWdlbnRDb21tYW5kMAESTAoJQ2xhaW1UYXNrEh4udGFza2d1aWxkLnYxLkNsYWltVGFza1JlcXVlc3QaHy50YXNrZ3VpbGQudjEuQ2xhaW1UYXNrUmVzcG9uc2USYQoQUmVwb3J0VGFza1Jlc3VsdBIlLnRhc2tndWlsZC52MS5SZXBvcnRUYXNrUmVzdWx0UmVxdWVzdBomLnRhc2tndWlsZC52MS5SZXBvcnRUYXNrUmVzdWx0UmVzcG9uc2USZAoRUmVwb3J0QWdlbnRTdGF0dXMSJi50YXNrZ3VpbGQudjEuUmVwb3J0QWdlbnRTdGF0dXNSZXF1ZXN0GicudGFza2d1aWxkLnYxLlJlcG9ydEFnZW50U3RhdHVzUmVzcG9uc2USTAoJSGVhcnRiZWF0Eh4udGFza2d1aWxkLnYxLkhlYXJ0YmVhdFJlcXVlc3QaHy50YXNrZ3VpbGQudjEuSGVhcnRiZWF0UmVzcG9uc2USZAoRQ3JlYXRlSW50ZXJhY3Rpb24SJi50YXNrZ3VpbGQudjEuQ3JlYXRlSW50ZXJhY3Rpb25SZXF1ZXN0GicudGFza2d1aWxkLnYxLkNyZWF0ZUludGVyYWN0aW9uUmVzcG9uc2UScwoWR2V0SW50ZXJhY3Rpb25SZXNwb25zZRIrLnRhc2tndWlsZC52MS5HZXRJbnRlcmFjdGlvblJlc3BvbnNlUmVxdWVzdBosLnRhc2tndWlsZC52MS5HZXRJbnRlcmFjdGlvblJlc3BvbnNlUmVzcG9uc2USTwoKU3luY0FnZW50cxIfLnRhc2tndWlsZC52MS5TeW5jQWdlbnRzUmVxdWVzdBogLnRhc2tndWlsZC52MS5TeW5jQWdlbnRzUmVzcG9uc2USWAoNUmVwb3J0VGFza0xvZxIiLnRhc2tndWlsZC52MS5SZXBvcnRUYXNrTG9nUmVxdWVzdBojLnRhc2tndWlsZC52MS5SZXBvcnRUYXNrTG9nUmVzcG9uc2VCugEKEGNvbS50YXNrZ3VpbGQudjFCEUFnZW50TWFuYWdlclByb3RvUAFaQmdpdGh1Yi5jb20va2F6ejE4Ny90YXNrZ3VpbGQvcHJvdG8vZ2VuL2dvL3Rhc2tndWlsZC92MTt0YXNrZ3VpbGR2MaICA1RYWKoCDFRhc2tndWlsZC5WMcoCDFRhc2tndWlsZFxWMeICGFRhc2tndWlsZFxWMVxHUEJNZXRhZGF0YeoCDVRhc2tndWlsZDo6VjFiBnByb3RvMw", [file_google_protobuf_timestamp, file_taskguild_v1_agent, file_taskguild_v1_interaction, file_taskguild_v1_task_log]);
+  fileDesc("CiB0YXNrZ3VpbGQvdjEvYWdlbnRfbWFuYWdlci5wcm90bxIMdGFza2d1aWxkLnYxImwKHEFnZW50TWFuYWdlclN1YnNjcmliZVJlcXVlc3QSGAoQYWdlbnRfbWFuYWdlcl9pZBgBIAEoCRIUCgxwcm9qZWN0X25hbWUYAiABKAkSHAoUbWF4X2NvbmN1cnJlbnRfdGFza3MYAyABKAUiyQIKDEFnZW50Q29tbWFuZBI8Cg50YXNrX2F2YWlsYWJsZRgBIAEoCzIiLnRhc2tndWlsZC52MS5UYXNrQXZhaWxhYmxlQ29tbWFuZEgAEjYKC2Fzc2lnbl90YXNrGAIgASgLMh8udGFza2d1aWxkLnYxLkFzc2lnblRhc2tDb21tYW5kSAASNgoLY2FuY2VsX3Rhc2sYAyABKAsyHy50YXNrZ3VpbGQudjEuQ2FuY2VsVGFza0NvbW1hbmRIABJIChRpbnRlcmFjdGlvbl9yZXNwb25zZRgEIAEoCzIoLnRhc2tndWlsZC52MS5JbnRlcmFjdGlvblJlc3BvbnNlQ29tbWFuZEgAEjYKC3N5bmNfYWdlbnRzGAUgASgLMh8udGFza2d1aWxkLnYxLlN5bmNBZ2VudHNDb21tYW5kSABCCQoHY29tbWFuZCLEAQoUVGFza0F2YWlsYWJsZUNvbW1hbmQSDwoHdGFza19pZBgBIAEoCRINCgV0aXRsZRgCIAEoCRIXCg9hZ2VudF9jb25maWdfaWQYAyABKAkSQgoIbWV0YWRhdGEYBCADKAsyMC50YXNrZ3VpbGQudjEuVGFza0F2YWlsYWJsZUNvbW1hbmQuTWV0YWRhdGFFbnRyeRovCg1NZXRhZGF0YUVudHJ5EgsKA2tleRgBIAEoCRINCgV2YWx1ZRgCIAEoCToCOAEi3gEKEUFzc2lnblRhc2tDb21tYW5kEg8KB3Rhc2tfaWQYASABKAkSFwoPYWdlbnRfY29uZmlnX2lkGAIgASgJEhQKDGluc3RydWN0aW9ucxgDIAEoCRIXCg93b3JrdHJlZV9icmFuY2gYBCABKAkSPwoIbWV0YWRhdGEYBSADKAsyLS50YXNrZ3VpbGQudjEuQXNzaWduVGFza0NvbW1hbmQuTWV0YWRhdGFFbnRyeRovCg1NZXRhZGF0YUVudHJ5EgsKA2tleRgBIAEoCRINCgV2YWx1ZRgCIAEoCToCOAEiNAoRQ2FuY2VsVGFza0NvbW1hbmQSDwoHdGFza19pZBgBIAEoCRIOCgZyZWFzb24YAiABKAkiRgoaSW50ZXJhY3Rpb25SZXNwb25zZUNvbW1hbmQSFgoOaW50ZXJhY3Rpb25faWQYASABKAkSEAoIcmVzcG9uc2UYAiABKAkiEwoRU3luY0FnZW50c0NvbW1hbmQiPQoQQ2xhaW1UYXNrUmVxdWVzdBIPCgd0YXNrX2lkGAEgASgJEhgKEGFnZW50X21hbmFnZXJfaWQYAiABKAkixQEKEUNsYWltVGFza1Jlc3BvbnNlEg8KB3N1Y2Nlc3MYASABKAgSFwoPYWdlbnRfY29uZmlnX2lkGAIgASgJEhQKDGluc3RydWN0aW9ucxgDIAEoCRI/CghtZXRhZGF0YRgEIAMoCzItLnRhc2tndWlsZC52MS5DbGFpbVRhc2tSZXNwb25zZS5NZXRhZGF0YUVudHJ5Gi8KDU1ldGFkYXRhRW50cnkSCwoDa2V5GAEgASgJEg0KBXZhbHVlGAIgASgJOgI4ASJgChdSZXBvcnRUYXNrUmVzdWx0UmVxdWVzdBIPCgd0YXNrX2lkGAEgASgJEg8KB3N1bW1hcnkYAyABKAkSFQoNZXJyb3JfbWVzc2FnZRgEIAEoCUoECAIQA1IGc3RhdHVzIhoKGFJlcG9ydFRhc2tSZXN1bHRSZXNwb25zZSKBAQoYUmVwb3J0QWdlbnRTdGF0dXNSZXF1ZXN0EhgKEGFnZW50X21hbmFnZXJfaWQYASABKAkSDwoHdGFza19pZBgCIAEoCRIpCgZzdGF0dXMYAyABKA4yGS50YXNrZ3VpbGQudjEuQWdlbnRTdGF0dXMSDwoHbWVzc2FnZRgEIAEoCSIbChlSZXBvcnRBZ2VudFN0YXR1c1Jlc3BvbnNlInEKEEhlYXJ0YmVhdFJlcXVlc3QSGAoQYWdlbnRfbWFuYWdlcl9pZBgBIAEoCRIUCgxhY3RpdmVfdGFza3MYAiABKAUSLQoJdGltZXN0YW1wGAMgASgLMhouZ29vZ2xlLnByb3RvYnVmLlRpbWVzdGFtcCITChFIZWFydGJlYXRSZXNwb25zZSLAAQoYQ3JlYXRlSW50ZXJhY3Rpb25SZXF1ZXN0Eg8KB3Rhc2tfaWQYASABKAkSEAoIYWdlbnRfaWQYAiABKAkSKwoEdHlwZRgDIAEoDjIdLnRhc2tndWlsZC52MS5JbnRlcmFjdGlvblR5cGUSDQoFdGl0bGUYBCABKAkSEwoLZGVzY3JpcHRpb24YBSABKAkSMAoHb3B0aW9ucxgGIAMoCzIfLnRhc2tndWlsZC52MS5JbnRlcmFjdGlvbk9wdGlvbiJLChlDcmVhdGVJbnRlcmFjdGlvblJlc3BvbnNlEi4KC2ludGVyYWN0aW9uGAEgASgLMhkudGFza2d1aWxkLnYxLkludGVyYWN0aW9uIjcKHUdldEludGVyYWN0aW9uUmVzcG9uc2VSZXF1ZXN0EhYKDmludGVyYWN0aW9uX2lkGAEgASgJIlAKHkdldEludGVyYWN0aW9uUmVzcG9uc2VSZXNwb25zZRIuCgtpbnRlcmFjdGlvbhgBIAEoCzIZLnRhc2tndWlsZC52MS5JbnRlcmFjdGlvbiIpChFTeW5jQWdlbnRzUmVxdWVzdBIUCgxwcm9qZWN0X25hbWUYASABKAkiQwoSU3luY0FnZW50c1Jlc3BvbnNlEi0KBmFnZW50cxgBIAMoCzIdLnRhc2tndWlsZC52MS5BZ2VudERlZmluaXRpb24iiQIKFFJlcG9ydFRhc2tMb2dSZXF1ZXN0Eg8KB3Rhc2tfaWQYASABKAkSKQoFbGV2ZWwYAiABKA4yGi50YXNrZ3VpbGQudjEuVGFza0xvZ0xldmVsEi8KCGNhdGVnb3J5GAMgASgOMh0udGFza2d1aWxkLnYxLlRhc2tMb2dDYXRlZ29yeRIPCgdtZXNzYWdlGAQgASgJEkIKCG1ldGFkYXRhGAUgAygLMjAudGFza2d1aWxkLnYxLlJlcG9ydFRhc2tMb2dSZXF1ZXN0Lk1ldGFkYXRhRW50cnkaLwoNTWV0YWRhdGFFbnRyeRILCgNrZXkYASABKAkSDQoFdmFsdWUYAiABKAk6AjgBIhcKFVJlcG9ydFRhc2tMb2dSZXNwb25zZSqOAQoLQWdlbnRTdGF0dXMSHAoYQUdFTlRfU1RBVFVTX1VOU1BFQ0lGSUVEEAASFQoRQUdFTlRfU1RBVFVTX0lETEUQARIYChRBR0VOVF9TVEFUVVNfUlVOTklORxACEhgKFEFHRU5UX1NUQVRVU19XQUlUSU5HEAMSFgoSQUdFTlRfU1RBVFVTX0VSUk9SEAQy1wYKE0FnZW50TWFuYWdlclNlcnZpY2USVQoJU3Vic2NyaWJlEioudGFza2d1aWxkLnYxLkFnZW50TWFuYWdlclN1YnNjcmliZVJlcXVlc3QaGi50YXNrZ3VpbGQudjEuQWdlbnRDb21tYW5kMAESTAoJQ2xhaW1UYXNrEh4udGFza2d1aWxkLnYxLkNsYWltVGFza1JlcXVlc3QaHy50YXNrZ3VpbGQudjEuQ2xhaW1UYXNrUmVzcG9uc2USYQoQUmVwb3J0VGFza1Jlc3VsdBIlLnRhc2tndWlsZC52MS5SZXBvcnRUYXNrUmVzdWx0UmVxdWVzdBomLnRhc2tndWlsZC52MS5SZXBvcnRUYXNrUmVzdWx0UmVzcG9uc2USZAoRUmVwb3J0QWdlbnRTdGF0dXMSJi50YXNrZ3VpbGQudjEuUmVwb3J0QWdlbnRTdGF0dXNSZXF1ZXN0GicudGFza2d1aWxkLnYxLlJlcG9ydEFnZW50U3RhdHVzUmVzcG9uc2USTAoJSGVhcnRiZWF0Eh4udGFza2d1aWxkLnYxLkhlYXJ0YmVhdFJlcXVlc3QaHy50YXNrZ3VpbGQudjEuSGVhcnRiZWF0UmVzcG9uc2USZAoRQ3JlYXRlSW50ZXJhY3Rpb24SJi50YXNrZ3VpbGQudjEuQ3JlYXRlSW50ZXJhY3Rpb25SZXF1ZXN0GicudGFza2d1aWxkLnYxLkNyZWF0ZUludGVyYWN0aW9uUmVzcG9uc2UScwoWR2V0SW50ZXJhY3Rpb25SZXNwb25zZRIrLnRhc2tndWlsZC52MS5HZXRJbnRlcmFjdGlvblJlc3BvbnNlUmVxdWVzdBosLnRhc2tndWlsZC52MS5HZXRJbnRlcmFjdGlvblJlc3BvbnNlUmVzcG9uc2USTwoKU3luY0FnZW50cxIfLnRhc2tndWlsZC52MS5TeW5jQWdlbnRzUmVxdWVzdBogLnRhc2tndWlsZC52MS5TeW5jQWdlbnRzUmVzcG9uc2USWAoNUmVwb3J0VGFza0xvZxIiLnRhc2tndWlsZC52MS5SZXBvcnRUYXNrTG9nUmVxdWVzdBojLnRhc2tndWlsZC52MS5SZXBvcnRUYXNrTG9nUmVzcG9uc2VCvwEKEGNvbS50YXNrZ3VpbGQudjFCEUFnZW50TWFuYWdlclByb3RvUAFaR2dpdGh1Yi5jb20va2F6ejE4Ny90YXNrZ3VpbGQvYmFja2VuZC9nZW4vcHJvdG8vdGFza2d1aWxkL3YxO3Rhc2tndWlsZHYxogIDVFhYqgIMVGFza2d1aWxkLlYxygIMVGFza2d1aWxkXFYx4gIYVGFza2d1aWxkXFYxXEdQQk1ldGFkYXRh6gINVGFza2d1aWxkOjpWMWIGcHJvdG8z", [file_google_protobuf_timestamp, file_taskguild_v1_agent, file_taskguild_v1_interaction, file_taskguild_v1_task_log]);
 
 /**
  * @generated from message taskguild.v1.AgentManagerSubscribeRequest
