+++ proto/gen/ts/taskguild/v1/agent_pb.ts
@@ -14,7 +14,7 @@ import type { Message } from "@bufbuild/protobuf";
  * Describes the file taskguild/v1/agent.proto.
  */
 export const file_taskguild_v1_agent: GenFile = /*@__PURE__*/
-  fileDesc("Chh0YXNrZ3VpbGQvdjEvYWdlbnQucHJvdG8SDHRhc2tndWlsZC52MSLIAgoPQWdlbnREZWZpbml0aW9uEgoKAmlkGAEgASgJEhIKCnByb2plY3RfaWQYAiABKAkSDAoEbmFtZRgDIAEoCRITCgtkZXNjcmlwdGlvbhgEIAEoCRIOCgZwcm9tcHQYBSABKAkSDQoFdG9vbHMYBiADKAkSGAoQZGlzYWxsb3dlZF90b29scxgHIAMoCRINCgVtb2RlbBgIIAEoCRIXCg9wZXJtaXNzaW9uX21vZGUYCSABKAkSDgoGc2tpbGxzGAogAygJEg4KBm1lbW9yeRgLIAEoCRIRCglpc19zeW5jZWQYDCABKAgSLgoKY3JlYXRlZF9hdBgNIAEoCzIaLmdvb2dsZS5wcm90b2J1Zi5UaW1lc3RhbXASLgoKdXBkYXRlZF9hdBgOIAEoCzIaLmdvb2dsZS5wcm90b2J1Zi5UaW1lc3RhbXAizAEKEkNyZWF0ZUFnZW50UmVxdWVzdBISCgpwcm9qZWN0X2lkGAEgASgJEgwKBG5hbWUYAiABKAkSEwoLZGVzY3JpcHRpb24YAyABKAkSDgoGcHJvbXB0GAQgASgJEg0KBXRvb2xzGAUgAygJEhgKEGRpc2FsbG93ZWRfdG9vbHMYBiADKAkSDQoFbW9kZWwYByABKAkSFwoPcGVybWlzc2lvbl9tb2RlGAggASgJEg4KBnNraWxscxgJIAMoCRIOCgZtZW1vcnkYCiABKAkiQwoTQ3JlYXRlQWdlbnRSZXNwb25zZRIsCgVhZ2VudBgBIAEoCzIdLnRhc2tndWlsZC52MS5BZ2VudERlZmluaXRpb24iHQoPR2V0QWdlbnRSZXF1ZXN0EgoKAmlkGAEgASgJIkAKEEdldEFnZW50UmVzcG9uc2USLAoFYWdlbnQYASABKAsyHS50YXNrZ3VpbGQudjEuQWdlbnREZWZpbml0aW9uIlwKEUxpc3RBZ2VudHNSZXF1ZXN0EhIKCnByb2plY3RfaWQYASABKAkSMwoKcGFnaW5hdGlvbhgCIAEoCzIfLnRhc2tndWlsZC52MS5QYWdpbmF0aW9uUmVxdWVzdCJ5ChJMaXN0QWdlbnRzUmVzcG9uc2USLQoGYWdlbnRzGAEgAygLMh0udGFza2d1aWxkLnYxLkFnZW50RGVmaW5pdGlvbhI0CgpwYWdpbmF0aW9uGAIgASgLMiAudGFza2d1aWxkLnYxLlBhZ2luYXRpb25SZXNwb25zZSLEAQoSVXBkYXRlQWdlbnRSZXF1ZXN0EgoKAmlkGAEgASgJEgwKBG5hbWUYAiABKAkSEwoLZGVzY3JpcHRpb24YAyABKAkSDgoGcHJvbXB0GAQgASgJEg0KBXRvb2xzGAUgAygJEhgKEGRpc2FsbG93ZWRfdG9vbHMYBiADKAkSDQoFbW9kZWwYByABKAkSFwoPcGVybWlzc2lvbl9tb2RlGAggASgJEg4KBnNraWxscxgJIAMoCRIOCgZtZW1vcnkYCiABKAkiQwoTVXBkYXRlQWdlbnRSZXNwb25zZRIsCgVhZ2VudBgBIAEoCzIdLnRhc2tndWlsZC52MS5BZ2VudERlZmluaXRpb24iIAoSRGVsZXRlQWdlbnRSZXF1ZXN0EgoKAmlkGAEgASgJIhUKE0RlbGV0ZUFnZW50UmVzcG9uc2UiQQoYU3luY0FnZW50c0Zyb21EaXJSZXF1ZXN0EhIKCnByb2plY3RfaWQYASABKAkSEQoJZGlyZWN0b3J5GAIgASgJImwKGVN5bmNBZ2VudHNGcm9tRGlyUmVzcG9uc2USLQoGYWdlbnRzGAEgAygLMh0udGFza2d1aWxkLnYxLkFnZW50RGVmaW5pdGlvbhIPCgdjcmVhdGVkGAIgASgFEg8KB3VwZGF0ZWQYAyABKAUyjAQKDEFnZW50U2VydmljZRJSCgtDcmVhdGVBZ2VudBIgLnRhc2tndWlsZC52MS5DcmVhdGVBZ2VudFJlcXVlc3QaIS50YXNrZ3VpbGQudjEuQ3JlYXRlQWdlbnRSZXNwb25zZRJJCghHZXRBZ2VudBIdLnRhc2tndWlsZC52MS5HZXRBZ2VudFJlcXVlc3QaHi50YXNrZ3VpbGQudjEuR2V0QWdlbnRSZXNwb25zZRJPCgpMaXN0QWdlbnRzEh8udGFza2d1aWxkLnYxLkxpc3RBZ2VudHNSZXF1ZXN0GiAudGFza2d1aWxkLnYxLkxpc3RBZ2VudHNSZXNwb25zZRJSCgtVcGRhdGVBZ2VudBIgLnRhc2tndWlsZC52MS5VcGRhdGVBZ2VudFJlcXVlc3QaIS50YXNrZ3VpbGQudjEuVXBkYXRlQWdlbnRSZXNwb25zZRJSCgtEZWxldGVBZ2VudBIgLnRhc2tndWlsZC52MS5EZWxldGVBZ2VudFJlcXVlc3QaIS50YXNrZ3VpbGQudjEuRGVsZXRlQWdlbnRSZXNwb25zZRJkChFTeW5jQWdlbnRzRnJvbURpchImLnRhc2tndWlsZC52MS5TeW5jQWdlbnRzRnJvbURpclJlcXVlc3QaJy50YXNrZ3VpbGQudjEuU3luY0FnZW50c0Zyb21EaXJSZXNwb25zZUKzAQoQY29tLnRhc2tndWlsZC52MUIKQWdlbnRQcm90b1ABWkJnaXRodWIuY29tL2thenoxODcvdGFza2d1aWxkL3Byb3RvL2dlbi9nby90YXNrZ3VpbGQvdjE7dGFza2d1aWxkdjGiAgNUWFiqAgxUYXNrZ3VpbGQuVjHKAgxUYXNrZ3VpbGRcVjHiAhhUYXNrZ3VpbGRcVjFcR1BCTWV0YWRhdGHqAg1UYXNrZ3VpbGQ6OlYxYgZwcm90bzM", [file_google_protobuf_timestamp, file_taskguild_v1_common]);
+  fileDesc("Chh0YXNrZ3VpbGQvdjEvYWdlbnQucHJvdG8SDHRhc2tndWlsZC52MSLIAgoPQWdlbnREZWZpbml0aW9uEgoKAmlkGAEgASgJEhIKCnByb2plY3RfaWQYAiABKAkSDAoEbmFtZRgDIAEoCRITCgtkZXNjcmlwdGlvbhgEIAEoCRIOCgZwcm9tcHQYBSABKAkSDQoFdG9vbHMYBiADKAkSGAoQZGlzYWxsb3dlZF90b29scxgHIAMoCRINCgVtb2RlbBgIIAEoCRIXCg9wZXJtaXNzaW9uX21vZGUYCSABKAkSDgoGc2tpbGxzGAogAygJEg4KBm1lbW9yeRgLIAEoCRIRCglpc19zeW5jZWQYDCABKAgSLgoKY3JlYXRlZF9hdBgNIAEoCzIaLmdvb2dsZS5wcm90b2J1Zi5UaW1lc3RhbXASLgoKdXBkYXRlZF9hdBgOIAEoCzIaLmdvb2dsZS5wcm90b2J1Zi5UaW1lc3RhbXAizAEKEkNyZWF0ZUFnZW50UmVxdWVzdBISCgpwcm9qZWN0X2lkGAEgASgJEgwKBG5hbWUYAiABKAkSEwoLZGVzY3JpcHRpb24YAyABKAkSDgoGcHJvbXB0GAQgASgJEg0KBXRvb2xzGAUgAygJEhgKEGRpc2FsbG93ZWRfdG9vbHMYBiADKAkSDQoFbW9kZWwYByABKAkSFwoPcGVybWlzc2lvbl9tb2RlGAggASgJEg4KBnNraWxscxgJIAMoCRIOCgZtZW1vcnkYCiABKAkiQwoTQ3JlYXRlQWdlbnRSZXNwb25zZRIsCgVhZ2VudBgBIAEoCzIdLnRhc2tndWlsZC52MS5BZ2VudERlZmluaXRpb24iHQoPR2V0QWdlbnRSZXF1ZXN0EgoKAmlkGAEgASgJIkAKEEdldEFnZW50UmVzcG9uc2USLAoFYWdlbnQYASABKAsyHS50YXNrZ3VpbGQudjEuQWdlbnREZWZpbml0aW9uIlwKEUxpc3RBZ2VudHNSZXF1ZXN0EhIKCnByb2plY3RfaWQYASABKAkSMwoKcGFnaW5hdGlvbhgCIAEoCzIfLnRhc2tndWlsZC52MS5QYWdpbmF0aW9uUmVxdWVzdCJ5ChJMaXN0QWdlbnRzUmVzcG9uc2USLQoGYWdlbnRzGAEgAygLMh0udGFza2d1aWxkLnYxLkFnZW50RGVmaW5pdGlvbhI0CgpwYWdpbmF0aW9uGAIgASgLMiAudGFza2d1aWxkLnYxLlBhZ2luYXRpb25SZXNwb25zZSLEAQoSVXBkYXRlQWdlbnRSZXF1ZXN0EgoKAmlkGAEgASgJEgwKBG5hbWUYAiABKAkSEwoLZGVzY3JpcHRpb24YAyABKAkSDgoGcHJvbXB0GAQgASgJEg0KBXRvb2xzGAUgAygJEhgKEGRpc2FsbG93ZWRfdG9vbHMYBiADKAkSDQoFbW9kZWwYByABKAkSFwoPcGVybWlzc2lvbl9tb2RlGAggASgJEg4KBnNraWxscxgJIAMoCRIOCgZtZW1vcnkYCiABKAkiQwoTVXBkYXRlQWdlbnRSZXNwb25zZRIsCgVhZ2VudBgBIAEoCzIdLnRhc2tndWlsZC52MS5BZ2VudERlZmluaXRpb24iIAoSRGVsZXRlQWdlbnRSZXF1ZXN0EgoKAmlkGAEgASgJIhUKE0RlbGV0ZUFnZW50UmVzcG9uc2UiQQoYU3luY0FnZW50c0Zyb21EaXJSZXF1ZXN0EhIKCnByb2plY3RfaWQYASABKAkSEQoJZGlyZWN0b3J5GAIgASgJImwKGVN5bmNBZ2VudHNGcm9tRGlyUmVzcG9uc2USLQoGYWdlbnRzGAEgAygLMh0udGFza2d1aWxkLnYxLkFnZW50RGVmaW5pdGlvbhIPCgdjcmVhdGVkGAIgASgFEg8KB3VwZGF0ZWQYAyABKAUyjAQKDEFnZW50U2VydmljZRJSCgtDcmVhdGVBZ2VudBIgLnRhc2tndWlsZC52MS5DcmVhdGVBZ2VudFJlcXVlc3QaIS50YXNrZ3VpbGQudjEuQ3JlYXRlQWdlbnRSZXNwb25zZRJJCghHZXRBZ2VudBIdLnRhc2tndWlsZC52MS5HZXRBZ2VudFJlcXVlc3QaHi50YXNrZ3VpbGQudjEuR2V0QWdlbnRSZXNwb25zZRJPCgpMaXN0QWdlbnRzEh8udGFza2d1aWxkLnYxLkxpc3RBZ2VudHNSZXF1ZXN0GiAudGFza2d1aWxkLnYxLkxpc3RBZ2VudHNSZXNwb25zZRJSCgtVcGRhdGVBZ2VudBIgLnRhc2tndWlsZC52MS5VcGRhdGVBZ2VudFJlcXVlc3QaIS50YXNrZ3VpbGQudjEuVXBkYXRlQWdlbnRSZXNwb25zZRJSCgtEZWxldGVBZ2VudBIgLnRhc2tndWlsZC52MS5EZWxldGVBZ2VudFJlcXVlc3QaIS50YXNrZ3VpbGQudjEuRGVsZXRlQWdlbnRSZXNwb25zZRJkChFTeW5jQWdlbnRzRnJvbURpchImLnRhc2tndWlsZC52MS5TeW5jQWdlbnRzRnJvbURpclJlcXVlc3QaJy50YXNrZ3VpbGQudjEuU3luY0FnZW50c0Zyb21EaXJSZXNwb25zZUK4AQoQY29tLnRhc2tndWlsZC52MUIKQWdlbnRQcm90b1ABWkdnaXRodWIuY29tL2thenoxODcvdGFza2d1aWxkL2JhY2tlbmQvZ2VuL3Byb3RvL3Rhc2tndWlsZC92MTt0YXNrZ3VpbGR2MaICA1RYWKoCDFRhc2tndWlsZC5WMcoCDFRhc2tndWlsZFxWMeICGFRhc2tndWlsZFxWMVxHUEJNZXRhZGF0YeoCDVRhc2tndWlsZDo6VjFiBnByb3RvMw", [file_google_protobuf_timestamp, file_taskguild_v1_common]);
 
 /**
  * AgentDefinition defines a reusable agent that can be assigned to workflow statuses.
