+++ proto/gen/ts/taskguild/v1/interaction_pb.ts
@@ -14,7 +14,7 @@ import type { Message } from "@bufbuild/protobuf";
  * Describes the file taskguild/v1/interaction.proto.
  */
 export const file_taskguild_v1_interaction: GenFile = /*@__PURE__*/
-  fileDesc("Ch50YXNrZ3VpbGQvdjEvaW50ZXJhY3Rpb24ucHJvdG8SDHRhc2tndWlsZC52MSLkAgoLSW50ZXJhY3Rpb24SCgoCaWQYASABKAkSDwoHdGFza19pZBgCIAEoCRIQCghhZ2VudF9pZBgDIAEoCRIrCgR0eXBlGAQgASgOMh0udGFza2d1aWxkLnYxLkludGVyYWN0aW9uVHlwZRIvCgZzdGF0dXMYBSABKA4yHy50YXNrZ3VpbGQudjEuSW50ZXJhY3Rpb25TdGF0dXMSDQoFdGl0bGUYBiABKAkSEwoLZGVzY3JpcHRpb24YByABKAkSMAoHb3B0aW9ucxgIIAMoCzIfLnRhc2tndWlsZC52MS5JbnRlcmFjdGlvbk9wdGlvbhIQCghyZXNwb25zZRgJIAEoCRIuCgpjcmVhdGVkX2F0GAogASgLMhouZ29vZ2xlLnByb3RvYnVmLlRpbWVzdGFtcBIwCgxyZXNwb25kZWRfYXQYCyABKAsyGi5nb29nbGUucHJvdG9idWYuVGltZXN0YW1wIkYKEUludGVyYWN0aW9uT3B0aW9uEg0KBWxhYmVsGAEgASgJEg0KBXZhbHVlGAIgASgJEhMKC2Rlc2NyaXB0aW9uGAMgASgJIkIKEEludGVyYWN0aW9uRXZlbnQSLgoLaW50ZXJhY3Rpb24YASABKAsyGS50YXNrZ3VpbGQudjEuSW50ZXJhY3Rpb24iqwEKF0xpc3RJbnRlcmFjdGlvbnNSZXF1ZXN0EhIKCnByb2plY3RfaWQYASABKAkSDwoHdGFza19pZBgCIAEoCRI2Cg1zdGF0dXNfZmlsdGVyGAMgASgOMh8udGFza2d1aWxkLnYxLkludGVyYWN0aW9uU3RhdHVzEjMKCnBhZ2luYXRpb24YBCABKAsyHy50YXNrZ3VpbGQudjEuUGFnaW5hdGlvblJlcXVlc3QigQEKGExpc3RJbnRlcmFjdGlvbnNSZXNwb25zZRIvCgxpbnRlcmFjdGlvbnMYASADKAsyGS50YXNrZ3VpbGQudjEuSW50ZXJhY3Rpb24SNAoKcGFnaW5hdGlvbhgCIAEoCzIgLnRhc2tndWlsZC52MS5QYWdpbmF0aW9uUmVzcG9uc2UiOwobUmVzcG9uZFRvSW50ZXJhY3Rpb25SZXF1ZXN0EgoKAmlkGAEgASgJEhAKCHJlc3BvbnNlGAIgASgJIk4KHFJlc3BvbmRUb0ludGVyYWN0aW9uUmVzcG9uc2USLgoLaW50ZXJhY3Rpb24YASABKAsyGS50YXNrZ3VpbGQudjEuSW50ZXJhY3Rpb24iNgoSU2VuZE1lc3NhZ2VSZXF1ZXN0Eg8KB3Rhc2tfaWQYASABKAkSDwoHbWVzc2FnZRgCIAEoCSJFChNTZW5kTWVzc2FnZVJlc3BvbnNlEi4KC2ludGVyYWN0aW9uGAEgASgLMhkudGFza2d1aWxkLnYxLkludGVyYWN0aW9uIi8KHFN1YnNjcmliZUludGVyYWN0aW9uc1JlcXVlc3QSDwoHdGFza19pZBgBIAEoCSrBAQoPSW50ZXJhY3Rpb25UeXBlEiAKHElOVEVSQUNUSU9OX1RZUEVfVU5TUEVDSUZJRUQQABInCiNJTlRFUkFDVElPTl9UWVBFX1BFUk1JU1NJT05fUkVRVUVTVBABEh0KGUlOVEVSQUNUSU9OX1RZUEVfUVVFU1RJT04QAhIhCh1JTlRFUkFDVElPTl9UWVBFX05PVElGSUNBVElPThADEiEKHUlOVEVSQUNUSU9OX1RZUEVfVVNFUl9NRVNTQUdFEAQqmQEKEUludGVyYWN0aW9uU3RhdHVzEiIKHklOVEVSQUNUSU9OX1NUQVRVU19VTlNQRUNJRklFRBAAEh4KGklOVEVSQUNUSU9OX1NUQVRVU19QRU5ESU5HEAESIAocSU5URVJBQ1RJT05fU1RBVFVTX1JFU1BPTkRFRBACEh4KGklOVEVSQUNUSU9OX1NUQVRVU19FWFBJUkVEEAMyoQMKEkludGVyYWN0aW9uU2VydmljZRJhChBMaXN0SW50ZXJhY3Rpb25zEiUudGFza2d1aWxkLnYxLkxpc3RJbnRlcmFjdGlvbnNSZXF1ZXN0GiYudGFza2d1aWxkLnYxLkxpc3RJbnRlcmFjdGlvbnNSZXNwb25zZRJtChRSZXNwb25kVG9JbnRlcmFjdGlvbhIpLnRhc2tndWlsZC52MS5SZXNwb25kVG9JbnRlcmFjdGlvblJlcXVlc3QaKi50YXNrZ3VpbGQudjEuUmVzcG9uZFRvSW50ZXJhY3Rpb25SZXNwb25zZRJSCgtTZW5kTWVzc2FnZRIgLnRhc2tndWlsZC52MS5TZW5kTWVzc2FnZVJlcXVlc3QaIS50YXNrZ3VpbGQudjEuU2VuZE1lc3NhZ2VSZXNwb25zZRJlChVTdWJzY3JpYmVJbnRlcmFjdGlvbnMSKi50YXNrZ3VpbGQudjEuU3Vic2NyaWJlSW50ZXJhY3Rpb25zUmVxdWVzdBoeLnRhc2tndWlsZC52MS5JbnRlcmFjdGlvbkV2ZW50MAFCuQEKEGNvbS50YXNrZ3VpbGQudjFCEEludGVyYWN0aW9uUHJvdG9QAVpCZ2l0aHViLmNvbS9rYXp6MTg3L3Rhc2tndWlsZC9wcm90by9nZW4vZ28vdGFza2d1aWxkL3YxO3Rhc2tndWlsZHYxogIDVFhYqgIMVGFza2d1aWxkLlYxygIMVGFza2d1aWxkXFYx4gIYVGFza2d1aWxkXFYxXEdQQk1ldGFkYXRh6gINVGFza2d1aWxkOjpWMWIGcHJvdG8z", [file_google_protobuf_timestamp, file_taskguild_v1_common]);
+  fileDesc("Ch50YXNrZ3VpbGQvdjEvaW50ZXJhY3Rpb24ucHJvdG8SDHRhc2tndWlsZC52MSLkAgoLSW50ZXJhY3Rpb24SCgoCaWQYASABKAkSDwoHdGFza19pZBgCIAEoCRIQCghhZ2VudF9pZBgDIAEoCRIrCgR0eXBlGAQgASgOMh0udGFza2d1aWxkLnYxLkludGVyYWN0aW9uVHlwZRIvCgZzdGF0dXMYBSABKA4yHy50YXNrZ3VpbGQudjEuSW50ZXJhY3Rpb25TdGF0dXMSDQoFdGl0bGUYBiABKAkSEwoLZGVzY3JpcHRpb24YByABKAkSMAoHb3B0aW9ucxgIIAMoCzIfLnRhc2tndWlsZC52MS5JbnRlcmFjdGlvbk9wdGlvbhIQCghyZXNwb25zZRgJIAEoCRIuCgpjcmVhdGVkX2F0GAogASgLMhouZ29vZ2xlLnByb3RvYnVmLlRpbWVzdGFtcBIwCgxyZXNwb25kZWRfYXQYCyABKAsyGi5nb29nbGUucHJvdG9idWYuVGltZXN0YW1wIkYKEUludGVyYWN0aW9uT3B0aW9uEg0KBWxhYmVsGAEgASgJEg0KBXZhbHVlGAIgASgJEhMKC2Rlc2NyaXB0aW9uGAMgASgJIkIKEEludGVyYWN0aW9uRXZlbnQSLgoLaW50ZXJhY3Rpb24YASABKAsyGS50YXNrZ3VpbGQudjEuSW50ZXJhY3Rpb24iqwEKF0xpc3RJbnRlcmFjdGlvbnNSZXF1ZXN0EhIKCnByb2plY3RfaWQYASABKAkSDwoHdGFza19pZBgCIAEoCRI2Cg1zdGF0dXNfZmlsdGVyGAMgASgOMh8udGFza2d1aWxkLnYxLkludGVyYWN0aW9uU3RhdHVzEjMKCnBhZ2luYXRpb24YBCABKAsyHy50YXNrZ3VpbGQudjEuUGFnaW5hdGlvblJlcXVlc3QigQEKGExpc3RJbnRlcmFjdGlvbnNSZXNwb25zZRIvCgxpbnRlcmFjdGlvbnMYASADKAsyGS50YXNrZ3VpbGQudjEuSW50ZXJhY3Rpb24SNAoKcGFnaW5hdGlvbhgCIAEoCzIgLnRhc2tndWlsZC52MS5QYWdpbmF0aW9uUmVzcG9uc2UiOwobUmVzcG9uZFRvSW50ZXJhY3Rpb25SZXF1ZXN0EgoKAmlkGAEgASgJEhAKCHJlc3BvbnNlGAIgASgJIk4KHFJlc3BvbmRUb0ludGVyYWN0aW9uUmVzcG9uc2USLgoLaW50ZXJhY3Rpb24YASABKAsyGS50YXNrZ3VpbGQudjEuSW50ZXJhY3Rpb24iNgoSU2VuZE1lc3NhZ2VSZXF1ZXN0Eg8KB3Rhc2tfaWQYASABKAkSDwoHbWVzc2FnZRgCIAEoCSJFChNTZW5kTWVzc2FnZVJlc3BvbnNlEi4KC2ludGVyYWN0aW9uGAEgASgLMhkudGFza2d1aWxkLnYxLkludGVyYWN0aW9uIi8KHFN1YnNjcmliZUludGVyYWN0aW9uc1JlcXVlc3QSDwoHdGFza19pZBgBIAEoCSrBAQoPSW50ZXJhY3Rpb25UeXBlEiAKHElOVEVSQUNUSU9OX1RZUEVfVU5TUEVDSUZJRUQQABInCiNJTlRFUkFDVElPTl9UWVBFX1BFUk1JU1NJT05fUkVRVUVTVBABEh0KGUlOVEVSQUNUSU9OX1RZUEVfUVVFU1RJT04QAhIhCh1JTlRFUkFDVElPTl9UWVBFX05PVElGSUNBVElPThADEiEKHUlOVEVSQUNUSU9OX1RZUEVfVVNFUl9NRVNTQUdFEAQqmQEKEUludGVyYWN0aW9uU3RhdHVzEiIKHklOVEVSQUNUSU9OX1NUQVRVU19VTlNQRUNJRklFRBAAEh4KGklOVEVSQUNUSU9OX1NUQVRVU19QRU5ESU5HEAESIAocSU5URVJBQ1RJT05fU1RBVFVTX1JFU1BPTkRFRBACEh4KGklOVEVSQUNUSU9OX1NUQVRVU19FWFBJUkVEEAMyoQMKEkludGVyYWN0aW9uU2VydmljZRJhChBMaXN0SW50ZXJhY3Rpb25zEiUudGFza2d1aWxkLnYxLkxpc3RJbnRlcmFjdGlvbnNSZXF1ZXN0GiYudGFza2d1aWxkLnYxLkxpc3RJbnRlcmFjdGlvbnNSZXNwb25zZRJtChRSZXNwb25kVG9JbnRlcmFjdGlvbhIpLnRhc2tndWlsZC52MS5SZXNwb25kVG9JbnRlcmFjdGlvblJlcXVlc3QaKi50YXNrZ3VpbGQudjEuUmVzcG9uZFRvSW50ZXJhY3Rpb25SZXNwb25zZRJSCgtTZW5kTWVzc2FnZRIgLnRhc2tndWlsZC52MS5TZW5kTWVzc2FnZVJlcXVlc3QaIS50YXNrZ3VpbGQudjEuU2VuZE1lc3NhZ2VSZXNwb25zZRJlChVTdWJzY3JpYmVJbnRlcmFjdGlvbnMSKi50YXNrZ3VpbGQudjEuU3Vic2NyaWJlSW50ZXJhY3Rpb25zUmVxdWVzdBoeLnRhc2tndWlsZC52MS5JbnRlcmFjdGlvbkV2ZW50MAFCvgEKEGNvbS50YXNrZ3VpbGQudjFCEEludGVyYWN0aW9uUHJvdG9QAVpHZ2l0aHViLmNvbS9rYXp6MTg3L3Rhc2tndWlsZC9iYWNrZW5kL2dlbi9wcm90by90YXNrZ3VpbGQvdjE7dGFza2d1aWxkdjGiAgNUWFiqAgxUYXNrZ3VpbGQuVjHKAgxUYXNrZ3VpbGRcVjHiAhhUYXNrZ3VpbGRcVjFcR1BCTWV0YWRhdGHqAg1UYXNrZ3VpbGQ6OlYxYgZwcm90bzM", [file_google_protobuf_timestamp, file_taskguild_v1_common]);
 
 /**
  * @generated from message taskguild.v1.Interaction
