+++ proto/gen/ts/taskguild/v1/push_notification_pb.ts
@@ -10,7 +10,7 @@ import type { Message } from "@bufbuild/protobuf";
  * Describes the file taskguild/v1/push_notification.proto.
  */
 export const file_taskguild_v1_push_notification: GenFile = /*@__PURE__*/
-  fileDesc("CiR0YXNrZ3VpbGQvdjEvcHVzaF9ub3RpZmljYXRpb24ucHJvdG8SDHRhc2tndWlsZC52MSIaChhHZXRWYXBpZFB1YmxpY0tleVJlcXVlc3QiLwoZR2V0VmFwaWRQdWJsaWNLZXlSZXNwb25zZRISCgpwdWJsaWNfa2V5GAEgASgJIlkKH1JlZ2lzdGVyUHVzaFN1YnNjcmlwdGlvblJlcXVlc3QSEAoIZW5kcG9pbnQYASABKAkSEgoKcDI1NmRoX2tleRgCIAEoCRIQCghhdXRoX2tleRgDIAEoCSIiCiBSZWdpc3RlclB1c2hTdWJzY3JpcHRpb25SZXNwb25zZSI1CiFVbnJlZ2lzdGVyUHVzaFN1YnNjcmlwdGlvblJlcXVlc3QSEAoIZW5kcG9pbnQYASABKAkiJAoiVW5yZWdpc3RlclB1c2hTdWJzY3JpcHRpb25SZXNwb25zZSIdChtTZW5kVGVzdE5vdGlmaWNhdGlvblJlcXVlc3QiHgocU2VuZFRlc3ROb3RpZmljYXRpb25SZXNwb25zZTLqAwoXUHVzaE5vdGlmaWNhdGlvblNlcnZpY2USZAoRR2V0VmFwaWRQdWJsaWNLZXkSJi50YXNrZ3VpbGQudjEuR2V0VmFwaWRQdWJsaWNLZXlSZXF1ZXN0GicudGFza2d1aWxkLnYxLkdldFZhcGlkUHVibGljS2V5UmVzcG9uc2USeQoYUmVnaXN0ZXJQdXNoU3Vic2NyaXB0aW9uEi0udGFza2d1aWxkLnYxLlJlZ2lzdGVyUHVzaFN1YnNjcmlwdGlvblJlcXVlc3QaLi50YXNrZ3VpbGQudjEuUmVnaXN0ZXJQdXNoU3Vic2NyaXB0aW9uUmVzcG9uc2USfwoaVW5yZWdpc3RlclB1c2hTdWJzY3JpcHRpb24SLy50YXNrZ3VpbGQudjEuVW5yZWdpc3RlclB1c2hTdWJzY3JpcHRpb25SZXF1ZXN0GjAudGFza2d1aWxkLnYxLlVucmVnaXN0ZXJQdXNoU3Vic2NyaXB0aW9uUmVzcG9uc2USbQoUU2VuZFRlc3ROb3RpZmljYXRpb24SKS50YXNrZ3VpbGQudjEuU2VuZFRlc3ROb3RpZmljYXRpb25SZXF1ZXN0GioudGFza2d1aWxkLnYxLlNlbmRUZXN0Tm90aWZpY2F0aW9uUmVzcG9uc2VCvgEKEGNvbS50YXNrZ3VpbGQudjFCFVB1c2hOb3RpZmljYXRpb25Qcm90b1ABWkJnaXRodWIuY29tL2thenoxODcvdGFza2d1aWxkL3Byb3RvL2dlbi9nby90YXNrZ3VpbGQvdjE7dGFza2d1aWxkdjGiAgNUWFiqAgxUYXNrZ3VpbGQuVjHKAgxUYXNrZ3VpbGRcVjHiAhhUYXNrZ3VpbGRcVjFcR1BCTWV0YWRhdGHqAg1UYXNrZ3VpbGQ6OlYxYgZwcm90bzM");
+  fileDesc("CiR0YXNrZ3VpbGQvdjEvcHVzaF9ub3RpZmljYXRpb24ucHJvdG8SDHRhc2tndWlsZC52MSIaChhHZXRWYXBpZFB1YmxpY0tleVJlcXVlc3QiLwoZR2V0VmFwaWRQdWJsaWNLZXlSZXNwb25zZRISCgpwdWJsaWNfa2V5GAEgASgJIlkKH1JlZ2lzdGVyUHVzaFN1YnNjcmlwdGlvblJlcXVlc3QSEAoIZW5kcG9pbnQYASABKAkSEgoKcDI1NmRoX2tleRgCIAEoCRIQCghhdXRoX2tleRgDIAEoCSIiCiBSZWdpc3RlclB1c2hTdWJzY3JpcHRpb25SZXNwb25zZSI1CiFVbnJlZ2lzdGVyUHVzaFN1YnNjcmlwdGlvblJlcXVlc3QSEAoIZW5kcG9pbnQYASABKAkiJAoiVW5yZWdpc3RlclB1c2hTdWJzY3JpcHRpb25SZXNwb25zZSIdChtTZW5kVGVzdE5vdGlmaWNhdGlvblJlcXVlc3QiHgocU2VuZFRlc3ROb3RpZmljYXRpb25SZXNwb25zZTLqAwoXUHVzaE5vdGlmaWNhdGlvblNlcnZpY2USZAoRR2V0VmFwaWRQdWJsaWNLZXkSJi50YXNrZ3VpbGQudjEuR2V0VmFwaWRQdWJsaWNLZXlSZXF1ZXN0GicudGFza2d1aWxkLnYxLkdldFZhcGlkUHVibGljS2V5UmVzcG9uc2USeQoYUmVnaXN0ZXJQdXNoU3Vic2NyaXB0aW9uEi0udGFza2d1aWxkLnYxLlJlZ2lzdGVyUHVzaFN1YnNjcmlwdGlvblJlcXVlc3QaLi50YXNrZ3VpbGQudjEuUmVnaXN0ZXJQdXNoU3Vic2NyaXB0aW9uUmVzcG9uc2USfwoaVW5yZWdpc3RlclB1c2hTdWJzY3JpcHRpb24SLy50YXNrZ3VpbGQudjEuVW5yZWdpc3RlclB1c2hTdWJzY3JpcHRpb25SZXF1ZXN0GjAudGFza2d1aWxkLnYxLlVucmVnaXN0ZXJQdXNoU3Vic2NyaXB0aW9uUmVzcG9uc2USbQoUU2VuZFRlc3ROb3RpZmljYXRpb24SKS50YXNrZ3VpbGQudjEuU2VuZFRlc3ROb3RpZmljYXRpb25SZXF1ZXN0GioudGFza2d1aWxkLnYxLlNlbmRUZXN0Tm90aWZpY2F0aW9uUmVzcG9uc2VCwwEKEGNvbS50YXNrZ3VpbGQudjFCFVB1c2hOb3RpZmljYXRpb25Qcm90b1ABWkdnaXRodWIuY29tL2thenoxODcvdGFza2d1aWxkL2JhY2tlbmQvZ2VuL3Byb3RvL3Rhc2tndWlsZC92MTt0YXNrZ3VpbGR2MaICA1RYWKoCDFRhc2tndWlsZC5WMcoCDFRhc2tndWlsZFxWMeICGFRhc2tndWlsZFxWMVxHUEJNZXRhZGF0YeoCDVRhc2tndWlsZDo6VjFiBnByb3RvMw");
 
 /**
  * @generated from message taskguild.v1.GetVapidPublicKeyRequest
