+++ proto/gen/ts/taskguild/v1/skill_pb.ts
@@ -14,7 +14,7 @@ import type { Message } from "@bufbuild/protobuf";
  * Describes the file taskguild/v1/skill.proto.
  */
 export const file_taskguild_v1_skill: GenFile = /*@__PURE__*/
-  fileDesc("Chh0YXNrZ3VpbGQvdjEvc2tpbGwucHJvdG8SDHRhc2tndWlsZC52MSLvAgoPU2tpbGxEZWZpbml0aW9uEgoKAmlkGAEgASgJEhIKCnByb2plY3RfaWQYAiABKAkSDAoEbmFtZRgDIAEoCRITCgtkZXNjcmlwdGlvbhgEIAEoCRIPCgdjb250ZW50GAUgASgJEiAKGGRpc2FibGVfbW9kZWxfaW52b2NhdGlvbhgGIAEoCBIWCg51c2VyX2ludm9jYWJsZRgHIAEoCBIVCg1hbGxvd2VkX3Rvb2xzGAggAygJEg0KBW1vZGVsGAkgASgJEg8KB2NvbnRleHQYCiABKAkSDQoFYWdlbnQYCyABKAkSFQoNYXJndW1lbnRfaGludBgMIAEoCRIRCglpc19zeW5jZWQYDSABKAgSLgoKY3JlYXRlZF9hdBgOIAEoCzIaLmdvb2dsZS5wcm90b2J1Zi5UaW1lc3RhbXASLgoKdXBkYXRlZF9hdBgPIAEoCzIaLmdvb2dsZS5wcm90b2J1Zi5UaW1lc3RhbXAi8wEKEkNyZWF0ZVNraWxsUmVxdWVzdBISCgpwcm9qZWN0X2lkGAEgASgJEgwKBG5hbWUYAiABKAkSEwoLZGVzY3JpcHRpb24YAyABKAkSDwoHY29udGVudBgEIAEoCRIgChhkaXNhYmxlX21vZGVsX2ludm9jYXRpb24YBSABKAgSFgoOdXNlcl9pbnZvY2FibGUYBiABKAgSFQoNYWxsb3dlZF90b29scxgHIAMoCRINCgVtb2RlbBgIIAEoCRIPCgdjb250ZXh0GAkgASgJEg0KBWFnZW50GAogASgJEhUKDWFyZ3VtZW50X2hpbnQYCyABKAkiQwoTQ3JlYXRlU2tpbGxSZXNwb25zZRIsCgVza2lsbBgBIAEoCzIdLnRhc2tndWlsZC52MS5Ta2lsbERlZmluaXRpb24iHQoPR2V0U2tpbGxSZXF1ZXN0EgoKAmlkGAEgASgJIkAKEEdldFNraWxsUmVzcG9uc2USLAoFc2tpbGwYASABKAsyHS50YXNrZ3VpbGQudjEuU2tpbGxEZWZpbml0aW9uIlwKEUxpc3RTa2lsbHNSZXF1ZXN0EhIKCnByb2plY3RfaWQYASABKAkSMwoKcGFnaW5hdGlvbhgCIAEoCzIfLnRhc2tndWlsZC52MS5QYWdpbmF0aW9uUmVxdWVzdCJ5ChJMaXN0U2tpbGxzUmVzcG9uc2USLQoGc2tpbGxzGAEgAygLMh0udGFza2d1aWxkLnYxLlNraWxsRGVmaW5pdGlvbhI0CgpwYWdpbmF0aW9uGAIgASgLMiAudGFza2d1aWxkLnYxLlBhZ2luYXRpb25SZXNwb25zZSLrAQoSVXBkYXRlU2tpbGxSZXF1ZXN0EgoKAmlkGAEgASgJEgwKBG5hbWUYAiABKAkSEwoLZGVzY3JpcHRpb24YAyABKAkSDwoHY29udGVudBgEIAEoCRIgChhkaXNhYmxlX21vZGVsX2ludm9jYXRpb24YBSABKAgSFgoOdXNlcl9pbnZvY2FibGUYBiABKAgSFQoNYWxsb3dlZF90b29scxgHIAMoCRINCgVtb2RlbBgIIAEoCRIPCgdjb250ZXh0GAkgASgJEg0KBWFnZW50GAogASgJEhUKDWFyZ3VtZW50X2hpbnQYCyABKAkiQwoTVXBkYXRlU2tpbGxSZXNwb25zZRIsCgVza2lsbBgBIAEoCzIdLnRhc2tndWlsZC52MS5Ta2lsbERlZmluaXRpb24iIAoSRGVsZXRlU2tpbGxSZXF1ZXN0EgoKAmlkGAEgASgJIhUKE0RlbGV0ZVNraWxsUmVzcG9uc2UiQQoYU3luY1NraWxsc0Zyb21EaXJSZXF1ZXN0EhIKCnByb2plY3RfaWQYASABKAkSEQoJZGlyZWN0b3J5GAIgASgJImwKGVN5bmNTa2lsbHNGcm9tRGlyUmVzcG9uc2USLQoGc2tpbGxzGAEgAygLMh0udGFza2d1aWxkLnYxLlNraWxsRGVmaW5pdGlvbhIPCgdjcmVhdGVkGAIgASgFEg8KB3VwZGF0ZWQYAyABKAUyjAQKDFNraWxsU2VydmljZRJSCgtDcmVhdGVTa2lsbBIgLnRhc2tndWlsZC52MS5DcmVhdGVTa2lsbFJlcXVlc3QaIS50YXNrZ3VpbGQudjEuQ3JlYXRlU2tpbGxSZXNwb25zZRJJCghHZXRTa2lsbBIdLnRhc2tndWlsZC52MS5HZXRTa2lsbFJlcXVlc3QaHi50YXNrZ3VpbGQudjEuR2V0U2tpbGxSZXNwb25zZRJPCgpMaXN0U2tpbGxzEh8udGFza2d1aWxkLnYxLkxpc3RTa2lsbHNSZXF1ZXN0GiAudGFza2d1aWxkLnYxLkxpc3RTa2lsbHNSZXNwb25zZRJSCgtVcGRhdGVTa2lsbBIgLnRhc2tndWlsZC52MS5VcGRhdGVTa2lsbFJlcXVlc3QaIS50YXNrZ3VpbGQudjEuVXBkYXRlU2tpbGxSZXNwb25zZRJSCgtEZWxldGVTa2lsbBIgLnRhc2tndWlsZC52MS5EZWxldGVTa2lsbFJlcXVlc3QaIS50YXNrZ3VpbGQudjEuRGVsZXRlU2tpbGxSZXNwb25zZRJkChFTeW5jU2tpbGxzRnJvbURpchImLnRhc2tndWlsZC52MS5TeW5jU2tpbGxzRnJvbURpclJlcXVlc3QaJy50YXNrZ3VpbGQudjEuU3luY1NraWxsc0Zyb21EaXJSZXNwb25zZUKzAQoQY29tLnRhc2tndWlsZC52MUIKU2tpbGxQcm90b1ABWkJnaXRodWIuY29tL2thenoxODcvdGFza2d1aWxkL3Byb3RvL2dlbi9nby90YXNrZ3VpbGQvdjE7dGFza2d1aWxkdjGiAgNUWFiqAgxUYXNrZ3VpbGQuVjHKAgxUYXNrZ3VpbGRcVjHiAhhUYXNrZ3VpbGRcVjFcR1BCTWV0YWRhdGHqAg1UYXNrZ3VpbGQ6OlYxYgZwcm90bzM", [file_google_protobuf_timestamp, file_taskguild_v1_common]);
+  fileDesc("Chh0YXNrZ3VpbGQvdjEvc2tpbGwucHJvdG8SDHRhc2tndWlsZC52MSLvAgoPU2tpbGxEZWZpbml0aW9uEgoKAmlkGAEgASgJEhIKCnByb2plY3RfaWQYAiABKAkSDAoEbmFtZRgDIAEoCRITCgtkZXNjcmlwdGlvbhgEIAEoCRIPCgdjb250ZW50GAUgASgJEiAKGGRpc2FibGVfbW9kZWxfaW52b2NhdGlvbhgGIAEoCBIWCg51c2VyX2ludm9jYWJsZRgHIAEoCBIVCg1hbGxvd2VkX3Rvb2xzGAggAygJEg0KBW1vZGVsGAkgASgJEg8KB2NvbnRleHQYCiABKAkSDQoFYWdlbnQYCyABKAkSFQoNYXJndW1lbnRfaGludBgMIAEoCRIRCglpc19zeW5jZWQYDSABKAgSLgoKY3JlYXRlZF9hdBgOIAEoCzIaLmdvb2dsZS5wcm90b2J1Zi5UaW1lc3RhbXASLgoKdXBkYXRlZF9hdBgPIAEoCzIaLmdvb2dsZS5wcm90b2J1Zi5UaW1lc3RhbXAi8wEKEkNyZWF0ZVNraWxsUmVxdWVzdBISCgpwcm9qZWN0X2lkGAEgASgJEgwKBG5hbWUYAiABKAkSEwoLZGVzY3JpcHRpb24YAyABKAkSDwoHY29udGVudBgEIAEoCRIgChhkaXNhYmxlX21vZGVsX2ludm9jYXRpb24YBSABKAgSFgoOdXNlcl9pbnZvY2FibGUYBiABKAgSFQoNYWxsb3dlZF90b29scxgHIAMoCRINCgVtb2RlbBgIIAEoCRIPCgdjb250ZXh0GAkgASgJEg0KBWFnZW50GAogASgJEhUKDWFyZ3VtZW50X2hpbnQYCyABKAkiQwoTQ3JlYXRlU2tpbGxSZXNwb25zZRIsCgVza2lsbBgBIAEoCzIdLnRhc2tndWlsZC52MS5Ta2lsbERlZmluaXRpb24iHQoPR2V0U2tpbGxSZXF1ZXN0EgoKAmlkGAEgASgJIkAKEEdldFNraWxsUmVzcG9uc2USLAoFc2tpbGwYASABKAsyHS50YXNrZ3VpbGQudjEuU2tpbGxEZWZpbml0aW9uIlwKEUxpc3RTa2lsbHNSZXF1ZXN0EhIKCnByb2plY3RfaWQYASABKAkSMwoKcGFnaW5hdGlvbhgCIAEoCzIfLnRhc2tndWlsZC52MS5QYWdpbmF0aW9uUmVxdWVzdCJ5ChJMaXN0U2tpbGxzUmVzcG9uc2USLQoGc2tpbGxzGAEgAygLMh0udGFza2d1aWxkLnYxLlNraWxsRGVmaW5pdGlvbhI0CgpwYWdpbmF0aW9uGAIgASgLMiAudGFza2d1aWxkLnYxLlBhZ2luYXRpb25SZXNwb25zZSLrAQoSVXBkYXRlU2tpbGxSZXF1ZXN0EgoKAmlkGAEgASgJEgwKBG5hbWUYAiABKAkSEwoLZGVzY3JpcHRpb24YAyABKAkSDwoHY29udGVudBgEIAEoCRIgChhkaXNhYmxlX21vZGVsX2ludm9jYXRpb24YBSABKAgSFgoOdXNlcl9pbnZvY2FibGUYBiABKAgSFQoNYWxsb3dlZF90b29scxgHIAMoCRINCgVtb2RlbBgIIAEoCRIPCgdjb250ZXh0GAkgASgJEg0KBWFnZW50GAogASgJEhUKDWFyZ3VtZW50X2hpbnQYCyABKAkiQwoTVXBkYXRlU2tpbGxSZXNwb25zZRIsCgVza2lsbBgBIAEoCzIdLnRhc2tndWlsZC52MS5Ta2lsbERlZmluaXRpb24iIAoSRGVsZXRlU2tpbGxSZXF1ZXN0EgoKAmlkGAEgASgJIhUKE0RlbGV0ZVNraWxsUmVzcG9uc2UiQQoYU3luY1NraWxsc0Zyb21EaXJSZXF1ZXN0EhIKCnByb2plY3RfaWQYASABKAkSEQoJZGlyZWN0b3J5GAIgASgJImwKGVN5bmNTa2lsbHNGcm9tRGlyUmVzcG9uc2USLQoGc2tpbGxzGAEgAygLMh0udGFza2d1aWxkLnYxLlNraWxsRGVmaW5pdGlvbhIPCgdjcmVhdGVkGAIgASgFEg8KB3VwZGF0ZWQYAyABKAUyjAQKDFNraWxsU2VydmljZRJSCgtDcmVhdGVTa2lsbBIgLnRhc2tndWlsZC52MS5DcmVhdGVTa2lsbFJlcXVlc3QaIS50YXNrZ3VpbGQudjEuQ3JlYXRlU2tpbGxSZXNwb25zZRJJCghHZXRTa2lsbBIdLnRhc2tndWlsZC52MS5HZXRTa2lsbFJlcXVlc3QaHi50YXNrZ3VpbGQudjEuR2V0U2tpbGxSZXNwb25zZRJPCgpMaXN0U2tpbGxzEh8udGFza2d1aWxkLnYxLkxpc3RTa2lsbHNSZXF1ZXN0GiAudGFza2d1aWxkLnYxLkxpc3RTa2lsbHNSZXNwb25zZRJSCgtVcGRhdGVTa2lsbBIgLnRhc2tndWlsZC52MS5VcGRhdGVTa2lsbFJlcXVlc3QaIS50YXNrZ3VpbGQudjEuVXBkYXRlU2tpbGxSZXNwb25zZRJSCgtEZWxldGVTa2lsbBIgLnRhc2tndWlsZC52MS5EZWxldGVTa2lsbFJlcXVlc3QaIS50YXNrZ3VpbGQudjEuRGVsZXRlU2tpbGxSZXNwb25zZRJkChFTeW5jU2tpbGxzRnJvbURpchImLnRhc2tndWlsZC52MS5TeW5jU2tpbGxzRnJvbURpclJlcXVlc3QaJy50YXNrZ3VpbGQudjEuU3luY1NraWxsc0Zyb21EaXJSZXNwb25zZUK4AQoQY29tLnRhc2tndWlsZC52MUIKU2tpbGxQcm90b1ABWkdnaXRodWIuY29tL2thenoxODcvdGFza2d1aWxkL2JhY2tlbmQvZ2VuL3Byb3RvL3Rhc2tndWlsZC92MTt0YXNrZ3VpbGR2MaICA1RYWKoCDFRhc2tndWlsZC5WMcoCDFRhc2tndWlsZFxWMeICGFRhc2tndWlsZFxWMVxHUEJNZXRhZGF0YeoCDVRhc2tndWlsZDo6VjFiBnByb3RvMw", [file_google_protobuf_timestamp, file_taskguild_v1_common]);
 
 /**
  * SkillDefinition defines a reusable skill (SKILL.md) that can be referenced by agents.
