+++ proto/gen/ts/taskguild/v1/task_log_pb.ts
@@ -14,7 +14,7 @@ import type { Message } from "@bufbuild/protobuf";
  * Describes the file taskguild/v1/task_log.proto.
  */
 export const file_taskguild_v1_task_log: GenFile = /*@__PURE__*/
-  fileDesc("Cht0YXNrZ3VpbGQvdjEvdGFza19sb2cucHJvdG8SDHRhc2tndWlsZC52MSKrAgoHVGFza0xvZxIKCgJpZBgBIAEoCRIPCgd0YXNrX2lkGAIgASgJEikKBWxldmVsGAMgASgOMhoudGFza2d1aWxkLnYxLlRhc2tMb2dMZXZlbBIvCghjYXRlZ29yeRgEIAEoDjIdLnRhc2tndWlsZC52MS5UYXNrTG9nQ2F0ZWdvcnkSDwoHbWVzc2FnZRgFIAEoCRI1CghtZXRhZGF0YRgGIAMoCzIjLnRhc2tndWlsZC52MS5UYXNrTG9nLk1ldGFkYXRhRW50cnkSLgoKY3JlYXRlZF9hdBgHIAEoCzIaLmdvb2dsZS5wcm90b2J1Zi5UaW1lc3RhbXAaLwoNTWV0YWRhdGFFbnRyeRILCgNrZXkYASABKAkSDQoFdmFsdWUYAiABKAk6AjgBIlsKE0xpc3RUYXNrTG9nc1JlcXVlc3QSDwoHdGFza19pZBgBIAEoCRIzCgpwYWdpbmF0aW9uGAIgASgLMh8udGFza2d1aWxkLnYxLlBhZ2luYXRpb25SZXF1ZXN0InEKFExpc3RUYXNrTG9nc1Jlc3BvbnNlEiMKBGxvZ3MYASADKAsyFS50YXNrZ3VpbGQudjEuVGFza0xvZxI0CgpwYWdpbmF0aW9uGAIgASgLMiAudGFza2d1aWxkLnYxLlBhZ2luYXRpb25SZXNwb25zZSqUAQoMVGFza0xvZ0xldmVsEh4KGlRBU0tfTE9HX0xFVkVMX1VOU1BFQ0lGSUVEEAASFwoTVEFTS19MT0dfTEVWRUxfSU5GTxABEhgKFFRBU0tfTE9HX0xFVkVMX0RFQlVHEAISFwoTVEFTS19MT0dfTEVWRUxfV0FSThADEhgKFFRBU0tfTE9HX0xFVkVMX0VSUk9SEAQqkAIKD1Rhc2tMb2dDYXRlZ29yeRIhCh1UQVNLX0xPR19DQVRFR09SWV9VTlNQRUNJRklFRBAAEiAKHFRBU0tfTE9HX0NBVEVHT1JZX1RVUk5fU1RBUlQQARIeChpUQVNLX0xPR19DQVRFR09SWV9UVVJOX0VORBACEiMKH1RBU0tfTE9HX0NBVEVHT1JZX1NUQVRVU19DSEFOR0UQAxIaChZUQVNLX0xPR19DQVRFR09SWV9IT09LEAQSHAoYVEFTS19MT0dfQ0FURUdPUllfU1RERVJSEAUSGwoXVEFTS19MT0dfQ0FURUdPUllfRVJST1IQBhIcChhUQVNLX0xPR19DQVRFR09SWV9TWVNURU0QBzJnCg5UYXNrTG9nU2VydmljZRJVCgxMaXN0VGFza0xvZ3MSIS50YXNrZ3VpbGQudjEuTGlzdFRhc2tMb2dzUmVxdWVzdBoiLnRhc2tndWlsZC52MS5MaXN0VGFza0xvZ3NSZXNwb25zZUK1AQoQY29tLnRhc2tndWlsZC52MUIMVGFza0xvZ1Byb3RvUAFaQmdpdGh1Yi5jb20va2F6ejE4Ny90YXNrZ3VpbGQvcHJvdG8vZ2VuL2dvL3Rhc2tndWlsZC92MTt0YXNrZ3VpbGR2MaICA1RYWKoCDFRhc2tndWlsZC5WMcoCDFRhc2tndWlsZFxWMeICGFRhc2tndWlsZFxWMVxHUEJNZXRhZGF0YeoCDVRhc2tndWlsZDo6VjFiBnByb3RvMw", [file_google_protobuf_timestamp, file_taskguild_v1_common]);
+  fileDesc("Cht0YXNrZ3VpbGQvdjEvdGFza19sb2cucHJvdG8SDHRhc2tndWlsZC52MSKrAgoHVGFza0xvZxIKCgJpZBgBIAEoCRIPCgd0YXNrX2lkGAIgASgJEikKBWxldmVsGAMgASgOMhoudGFza2d1aWxkLnYxLlRhc2tMb2dMZXZlbBIvCghjYXRlZ29yeRgEIAEoDjIdLnRhc2tndWlsZC52MS5UYXNrTG9nQ2F0ZWdvcnkSDwoHbWVzc2FnZRgFIAEoCRI1CghtZXRhZGF0YRgGIAMoCzIjLnRhc2tndWlsZC52MS5UYXNrTG9nLk1ldGFkYXRhRW50cnkSLgoKY3JlYXRlZF9hdBgHIAEoCzIaLmdvb2dsZS5wcm90b2J1Zi5UaW1lc3RhbXAaLwoNTWV0YWRhdGFFbnRyeRILCgNrZXkYASABKAkSDQoFdmFsdWUYAiABKAk6AjgBIlsKE0xpc3RUYXNrTG9nc1JlcXVlc3QSDwoHdGFza19pZBgBIAEoCRIzCgpwYWdpbmF0aW9uGAIgASgLMh8udGFza2d1aWxkLnYxLlBhZ2luYXRpb25SZXF1ZXN0InEKFExpc3RUYXNrTG9nc1Jlc3BvbnNlEiMKBGxvZ3MYASADKAsyFS50YXNrZ3VpbGQudjEuVGFza0xvZxI0CgpwYWdpbmF0aW9uGAIgASgLMiAudGFza2d1aWxkLnYxLlBhZ2luYXRpb25SZXNwb25zZSqUAQoMVGFza0xvZ0xldmVsEh4KGlRBU0tfTE9HX0xFVkVMX1VOU1BFQ0lGSUVEEAASFwoTVEFTS19MT0dfTEVWRUxfSU5GTxABEhgKFFRBU0tfTE9HX0xFVkVMX0RFQlVHEAISFwoTVEFTS19MT0dfTEVWRUxfV0FSThADEhgKFFRBU0tfTE9HX0xFVkVMX0VSUk9SEAQqkAIKD1Rhc2tMb2dDYXRlZ29yeRIhCh1UQVNLX0xPR19DQVRFR09SWV9VTlNQRUNJRklFRBAAEiAKHFRBU0tfTE9HX0NBVEVHT1JZX1RVUk5fU1RBUlQQARIeChpUQVNLX0xPR19DQVRFR09SWV9UVVJOX0VORBACEiMKH1RBU0tfTE9HX0NBVEVHT1JZX1NUQVRVU19DSEFOR0UQAxIaChZUQVNLX0xPR19DQVRFR09SWV9IT09LEAQSHAoYVEFTS19MT0dfQ0FURUdPUllfU1RERVJSEAUSGwoXVEFTS19MT0dfQ0FURUdPUllfRVJST1IQBhIcChhUQVNLX0xPR19DQVRFR09SWV9TWVNURU0QBzJnCg5UYXNrTG9nU2VydmljZRJVCgxMaXN0VGFza0xvZ3MSIS50YXNrZ3VpbGQudjEuTGlzdFRhc2tMb2dzUmVxdWVzdBoiLnRhc2tndWlsZC52MS5MaXN0VGFza0xvZ3NSZXNwb25zZUK6AQoQY29tLnRhc2tndWlsZC52MUIMVGFza0xvZ1Byb3RvUAFaR2dpdGh1Yi5jb20va2F6ejE4Ny90YXNrZ3VpbGQvYmFja2VuZC9nZW4vcHJvdG8vdGFza2d1aWxkL3YxO3Rhc2tndWlsZHYxogIDVFhYqgIMVGFza2d1aWxkLlYxygIMVGFza2d1aWxkXFYx4gIYVGFza2d1aWxkXFYxXEdQQk1ldGFkYXRh6gINVGFza2d1aWxkOjpWMWIGcHJvdG8z", [file_google_protobuf_timestamp, file_taskguild_v1_common]);
 
 /**
  * @generated from message taskguild.v1.TaskLog
