+++ proto/gen/ts/taskguild/v1/workflow_pb.ts
@@ -14,7 +14,7 @@ import type { Message } from "@bufbuild/protobuf";
  * Describes the file taskguild/v1/workflow.proto.
  */
 export const file_taskguild_v1_workflow: GenFile = /*@__PURE__*/
-  fileDesc("Cht0YXNrZ3VpbGQvdjEvd29ya2Zsb3cucHJvdG8SDHRhc2tndWlsZC52MSLOAgoIV29ya2Zsb3cSCgoCaWQYASABKAkSEgoKcHJvamVjdF9pZBgCIAEoCRIMCgRuYW1lGAMgASgJEhMKC2Rlc2NyaXB0aW9uGAQgASgJEi4KCHN0YXR1c2VzGAUgAygLMhwudGFza2d1aWxkLnYxLldvcmtmbG93U3RhdHVzEjAKDWFnZW50X2NvbmZpZ3MYBiADKAsyGS50YXNrZ3VpbGQudjEuQWdlbnRDb25maWcSLgoKY3JlYXRlZF9hdBgHIAEoCzIaLmdvb2dsZS5wcm90b2J1Zi5UaW1lc3RhbXASLgoKdXBkYXRlZF9hdBgIIAEoCzIaLmdvb2dsZS5wcm90b2J1Zi5UaW1lc3RhbXASHwoXZGVmYXVsdF9wZXJtaXNzaW9uX21vZGUYCSABKAkSHAoUZGVmYXVsdF91c2Vfd29ya3RyZWUYCiABKAgicwoKU3RhdHVzSG9vaxIKCgJpZBgBIAEoCRIQCghza2lsbF9pZBgCIAEoCRIqCgd0cmlnZ2VyGAMgASgOMhkudGFza2d1aWxkLnYxLkhvb2tUcmlnZ2VyEg0KBW9yZGVyGAQgASgFEgwKBG5hbWUYBSABKAkitQEKDldvcmtmbG93U3RhdHVzEgoKAmlkGAEgASgJEgwKBG5hbWUYAiABKAkSDQoFb3JkZXIYAyABKAUSEgoKaXNfaW5pdGlhbBgEIAEoCBITCgtpc190ZXJtaW5hbBgFIAEoCBIWCg50cmFuc2l0aW9uc190bxgGIAMoCRIQCghhZ2VudF9pZBgHIAEoCRInCgVob29rcxgIIAMoCzIYLnRhc2tndWlsZC52MS5TdGF0dXNIb29rIoUBCgtBZ2VudENvbmZpZxIKCgJpZBgBIAEoCRIaChJ3b3JrZmxvd19zdGF0dXNfaWQYAiABKAkSDAoEbmFtZRgDIAEoCRITCgtkZXNjcmlwdGlvbhgEIAEoCRIUCgxpbnN0cnVjdGlvbnMYBSABKAkSFQoNYWxsb3dlZF90b29scxgGIAMoCSLvAQoVQ3JlYXRlV29ya2Zsb3dSZXF1ZXN0EhIKCnByb2plY3RfaWQYASABKAkSDAoEbmFtZRgCIAEoCRITCgtkZXNjcmlwdGlvbhgDIAEoCRIuCghzdGF0dXNlcxgEIAMoCzIcLnRhc2tndWlsZC52MS5Xb3JrZmxvd1N0YXR1cxIwCg1hZ2VudF9jb25maWdzGAUgAygLMhkudGFza2d1aWxkLnYxLkFnZW50Q29uZmlnEh8KF2RlZmF1bHRfcGVybWlzc2lvbl9tb2RlGAYgASgJEhwKFGRlZmF1bHRfdXNlX3dvcmt0cmVlGAcgASgIIkIKFkNyZWF0ZVdvcmtmbG93UmVzcG9uc2USKAoId29ya2Zsb3cYASABKAsyFi50YXNrZ3VpbGQudjEuV29ya2Zsb3ciIAoSR2V0V29ya2Zsb3dSZXF1ZXN0EgoKAmlkGAEgASgJIj8KE0dldFdvcmtmbG93UmVzcG9uc2USKAoId29ya2Zsb3cYASABKAsyFi50YXNrZ3VpbGQudjEuV29ya2Zsb3ciXwoUTGlzdFdvcmtmbG93c1JlcXVlc3QSEgoKcHJvamVjdF9pZBgBIAEoCRIzCgpwYWdpbmF0aW9uGAIgASgLMh8udGFza2d1aWxkLnYxLlBhZ2luYXRpb25SZXF1ZXN0IngKFUxpc3RXb3JrZmxvd3NSZXNwb25zZRIpCgl3b3JrZmxvd3MYASADKAsyFi50YXNrZ3VpbGQudjEuV29ya2Zsb3cSNAoKcGFnaW5hdGlvbhgCIAEoCzIgLnRhc2tndWlsZC52MS5QYWdpbmF0aW9uUmVzcG9uc2Ui5wEKFVVwZGF0ZVdvcmtmbG93UmVxdWVzdBIKCgJpZBgBIAEoCRIMCgRuYW1lGAIgASgJEhMKC2Rlc2NyaXB0aW9uGAMgASgJEi4KCHN0YXR1c2VzGAQgAygLMhwudGFza2d1aWxkLnYxLldvcmtmbG93U3RhdHVzEjAKDWFnZW50X2NvbmZpZ3MYBSADKAsyGS50YXNrZ3VpbGQudjEuQWdlbnRDb25maWcSHwoXZGVmYXVsdF9wZXJtaXNzaW9uX21vZGUYBiABKAkSHAoUZGVmYXVsdF91c2Vfd29ya3RyZWUYByABKAgiQgoWVXBkYXRlV29ya2Zsb3dSZXNwb25zZRIoCgh3b3JrZmxvdxgBIAEoCzIWLnRhc2tndWlsZC52MS5Xb3JrZmxvdyIjChVEZWxldGVXb3JrZmxvd1JlcXVlc3QSCgoCaWQYASABKAkiGAoWRGVsZXRlV29ya2Zsb3dSZXNwb25zZSqkAQoLSG9va1RyaWdnZXISHAoYSE9PS19UUklHR0VSX1VOU1BFQ0lGSUVEEAASJgoiSE9PS19UUklHR0VSX0JFRk9SRV9UQVNLX0VYRUNVVElPThABEiUKIUhPT0tfVFJJR0dFUl9BRlRFUl9UQVNLX0VYRUNVVElPThACEigKJEhPT0tfVFJJR0dFUl9BRlRFUl9XT1JLVFJFRV9DUkVBVElPThADMtYDCg9Xb3JrZmxvd1NlcnZpY2USWwoOQ3JlYXRlV29ya2Zsb3cSIy50YXNrZ3VpbGQudjEuQ3JlYXRlV29ya2Zsb3dSZXF1ZXN0GiQudGFza2d1aWxkLnYxLkNyZWF0ZVdvcmtmbG93UmVzcG9uc2USUgoLR2V0V29ya2Zsb3cSIC50YXNrZ3VpbGQudjEuR2V0V29ya2Zsb3dSZXF1ZXN0GiEudGFza2d1aWxkLnYxLkdldFdvcmtmbG93UmVzcG9uc2USWAoNTGlzdFdvcmtmbG93cxIiLnRhc2tndWlsZC52MS5MaXN0V29ya2Zsb3dzUmVxdWVzdBojLnRhc2tndWlsZC52MS5MaXN0V29ya2Zsb3dzUmVzcG9uc2USWwoOVXBkYXRlV29ya2Zsb3cSIy50YXNrZ3VpbGQudjEuVXBkYXRlV29ya2Zsb3dSZXF1ZXN0GiQudGFza2d1aWxkLnYxLlVwZGF0ZVdvcmtmbG93UmVzcG9uc2USWwoORGVsZXRlV29ya2Zsb3cSIy50YXNrZ3VpbGQudjEuRGVsZXRlV29ya2Zsb3dSZXF1ZXN0GiQudGFza2d1aWxkLnYxLkRlbGV0ZVdvcmtmbG93UmVzcG9uc2VCtgEKEGNvbS50YXNrZ3VpbGQudjFCDVdvcmtmbG93UHJvdG9QAVpCZ2l0aHViLmNvbS9rYXp6MTg3L3Rhc2tndWlsZC9wcm90by9nZW4vZ28vdGFza2d1aWxkL3YxO3Rhc2tndWlsZHYxogIDVFhYqgIMVGFza2d1aWxkLlYxygIMVGFza2d1aWxkXFYx4gIYVGFza2d1aWxkXFYxXEdQQk1ldGFkYXRh6gINVGFza2d1aWxkOjpWMWIGcHJvdG8z", [file_google_protobuf_timestamp, file_taskguild_v1_common]);
+  fileDesc("Cht0YXNrZ3VpbGQvdjEvd29ya2Zsb3cucHJvdG8SDHRhc2tndWlsZC52MSLOAgoIV29ya2Zsb3cSCgoCaWQYASABKAkSEgoKcHJvamVjdF9pZBgCIAEoCRIMCgRuYW1lGAMgASgJEhMKC2Rlc2NyaXB0aW9uGAQgASgJEi4KCHN0YXR1c2VzGAUgAygLMhwudGFza2d1aWxkLnYxLldvcmtmbG93U3RhdHVzEjAKDWFnZW50X2NvbmZpZ3MYBiADKAsyGS50YXNrZ3VpbGQudjEuQWdlbnRDb25maWcSLgoKY3JlYXRlZF9hdBgHIAEoCzIaLmdvb2dsZS5wcm90b2J1Zi5UaW1lc3RhbXASLgoKdXBkYXRlZF9hdBgIIAEoCzIaLmdvb2dsZS5wcm90b2J1Zi5UaW1lc3RhbXASHwoXZGVmYXVsdF9wZXJtaXNzaW9uX21vZGUYCSABKAkSHAoUZGVmYXVsdF91c2Vfd29ya3RyZWUYCiABKAgicwoKU3RhdHVzSG9vaxIKCgJpZBgBIAEoCRIQCghza2lsbF9pZBgCIAEoCRIqCgd0cmlnZ2VyGAMgASgOMhkudGFza2d1aWxkLnYxLkhvb2tUcmlnZ2VyEg0KBW9yZGVyGAQgASgFEgwKBG5hbWUYBSABKAkitQEKDldvcmtmbG93U3RhdHVzEgoKAmlkGAEgASgJEgwKBG5hbWUYAiABKAkSDQoFb3JkZXIYAyABKAUSEgoKaXNfaW5pdGlhbBgEIAEoCBITCgtpc190ZXJtaW5hbBgFIAEoCBIWCg50cmFuc2l0aW9uc190bxgGIAMoCRIQCghhZ2VudF9pZBgHIAEoCRInCgVob29rcxgIIAMoCzIYLnRhc2tndWlsZC52MS5TdGF0dXNIb29rIoUBCgtBZ2VudENvbmZpZxIKCgJpZBgBIAEoCRIaChJ3b3JrZmxvd19zdGF0dXNfaWQYAiABKAkSDAoEbmFtZRgDIAEoCRITCgtkZXNjcmlwdGlvbhgEIAEoCRIUCgxpbnN0cnVjdGlvbnMYBSABKAkSFQoNYWxsb3dlZF90b29scxgGIAMoCSLvAQoVQ3JlYXRlV29ya2Zsb3dSZXF1ZXN0EhIKCnByb2plY3RfaWQYASABKAkSDAoEbmFtZRgCIAEoCRITCgtkZXNjcmlwdGlvbhgDIAEoCRIuCghzdGF0dXNlcxgEIAMoCzIcLnRhc2tndWlsZC52MS5Xb3JrZmxvd1N0YXR1cxIwCg1hZ2VudF9jb25maWdzGAUgAygLMhkudGFza2d1aWxkLnYxLkFnZW50Q29uZmlnEh8KF2RlZmF1bHRfcGVybWlzc2lvbl9tb2RlGAYgASgJEhwKFGRlZmF1bHRfdXNlX3dvcmt0cmVlGAcgASgIIkIKFkNyZWF0ZVdvcmtmbG93UmVzcG9uc2USKAoId29ya2Zsb3cYASABKAsyFi50YXNrZ3VpbGQudjEuV29ya2Zsb3ciIAoSR2V0V29ya2Zsb3dSZXF1ZXN0EgoKAmlkGAEgASgJIj8KE0dldFdvcmtmbG93UmVzcG9uc2USKAoId29ya2Zsb3cYASABKAsyFi50YXNrZ3VpbGQudjEuV29ya2Zsb3ciXwoUTGlzdFdvcmtmbG93c1JlcXVlc3QSEgoKcHJvamVjdF9pZBgBIAEoCRIzCgpwYWdpbmF0aW9uGAIgASgLMh8udGFza2d1aWxkLnYxLlBhZ2luYXRpb25SZXF1ZXN0IngKFUxpc3RXb3JrZmxvd3NSZXNwb25zZRIpCgl3b3JrZmxvd3MYASADKAsyFi50YXNrZ3VpbGQudjEuV29ya2Zsb3cSNAoKcGFnaW5hdGlvbhgCIAEoCzIgLnRhc2tndWlsZC52MS5QYWdpbmF0aW9uUmVzcG9uc2Ui5wEKFVVwZGF0ZVdvcmtmbG93UmVxdWVzdBIKCgJpZBgBIAEoCRIMCgRuYW1lGAIgASgJEhMKC2Rlc2NyaXB0aW9uGAMgASgJEi4KCHN0YXR1c2VzGAQgAygLMhwudGFza2d1aWxkLnYxLldvcmtmbG93U3RhdHVzEjAKDWFnZW50X2NvbmZpZ3MYBSADKAsyGS50YXNrZ3VpbGQudjEuQWdlbnRDb25maWcSHwoXZGVmYXVsdF9wZXJtaXNzaW9uX21vZGUYBiABKAkSHAoUZGVmYXVsdF91c2Vfd29ya3RyZWUYByABKAgiQgoWVXBkYXRlV29ya2Zsb3dSZXNwb25zZRIoCgh3b3JrZmxvdxgBIAEoCzIWLnRhc2tndWlsZC52MS5Xb3JrZmxvdyIjChVEZWxldGVXb3JrZmxvd1JlcXVlc3QSCgoCaWQYASABKAkiGAoWRGVsZXRlV29ya2Zsb3dSZXNwb25zZSqkAQoLSG9va1RyaWdnZXISHAoYSE9PS19UUklHR0VSX1VOU1BFQ0lGSUVEEAASJgoiSE9PS19UUklHR0VSX0JFRk9SRV9UQVNLX0VYRUNVVElPThABEiUKIUhPT0tfVFJJR0dFUl9BRlRFUl9UQVNLX0VYRUNVVElPThACEigKJEhPT0tfVFJJR0dFUl9BRlRFUl9XT1JLVFJFRV9DUkVBVElPThADMtYDCg9Xb3JrZmxvd1NlcnZpY2USWwoOQ3JlYXRlV29ya2Zsb3cSIy50YXNrZ3VpbGQudjEuQ3JlYXRlV29ya2Zsb3dSZXF1ZXN0GiQudGFza2d1aWxkLnYxLkNyZWF0ZVdvcmtmbG93UmVzcG9uc2USUgoLR2V0V29ya2Zsb3cSIC50YXNrZ3VpbGQudjEuR2V0V29ya2Zsb3dSZXF1ZXN0GiEudGFza2d1aWxkLnYxLkdldFdvcmtmbG93UmVzcG9uc2USWAoNTGlzdFdvcmtmbG93cxIiLnRhc2tndWlsZC52MS5MaXN0V29ya2Zsb3dzUmVxdWVzdBojLnRhc2tndWlsZC52MS5MaXN0V29ya2Zsb3dzUmVzcG9uc2USWwoOVXBkYXRlV29ya2Zsb3cSIy50YXNrZ3VpbGQudjEuVXBkYXRlV29ya2Zsb3dSZXF1ZXN0GiQudGFza2d1aWxkLnYxLlVwZGF0ZVdvcmtmbG93UmVzcG9uc2USWwoORGVsZXRlV29ya2Zsb3cSIy50YXNrZ3VpbGQudjEuRGVsZXRlV29ya2Zsb3dSZXF1ZXN0GiQudGFza2d1aWxkLnYxLkRlbGV0ZVdvcmtmbG93UmVzcG9uc2VCuwEKEGNvbS50YXNrZ3VpbGQudjFCDVdvcmtmbG93UHJvdG9QAVpHZ2l0aHViLmNvbS9rYXp6MTg3L3Rhc2tndWlsZC9iYWNrZW5kL2dlbi9wcm90by90YXNrZ3VpbGQvdjE7dGFza2d1aWxkdjGiAgNUWFiqAgxUYXNrZ3VpbGQuVjHKAgxUYXNrZ3VpbGRcVjHiAhhUYXNrZ3VpbGRcVjFcR1BCTWV0YWRhdGHqAg1UYXNrZ3VpbGQ6OlYxYgZwcm90bzM", [file_google_protobuf_timestamp, file_taskguild_v1_common]);
 
 /**
  * Workflow defines a project's task lifecycle with custom statuses and agent configurations.
